refactor(kit): split world mesher runtime
