+++ crates/aether-kit/src/world/mesher/mod.rs
@@ -28,11 +28,11 @@
 //!
 //! The ground is a single surface, world-space in meters (`1 cell = 1 m`),
 //! tiled exactly by its material regions. The cascade-resolved material
-//! grid runs through [`repartition`] with smoothing disabled (zero
+//! grid runs through [`contour::repartition`] with smoothing disabled (zero
 //! iterations — the crisp path, no chamfer), so the partition boundary is
 //! the raw marching-squares staircase. Wherever a cell and its one-sample
 //! surround are uniformly one material, the cell emits a flat keyed quilt
-//! cell in that material's [`flat_color`]; everywhere else the partition
+//! cell in that material's [`style::flat_color`]; everywhere else the partition
 //! marches per window, each label's polygon colored by its material keyed
 //! at the owning cell, saddles resolved by label order so every window
 //! tiles.
@@ -42,7 +42,7 @@
 //! Every vertex lifts onto the plate-resolved height surface
 //! ([`World::surface_height_in`]): cell heights blend into continuous
 //! slopes where neighbors sit within the step ceiling
-//! ([`STEP_MAX_OCTIMETERS`]) and break where they exceed it. Where a cell
+//! ([`crate::world::STEP_MAX_OCTIMETERS`]) and break where they exceed it. Where a cell
 //! carries authored per-point relief (`World::cell_has_height_relief`) the
 //! pass resolves one stride down: the interior cap tessellates to
 //! `SUB × SUB` subcell quads over point patches (`SubPatch`) so a subcell
@@ -71,57 +71,28 @@
 pub mod contour;
 pub mod style;
 
-use alloc::vec;
+mod constants;
+mod coverage;
+mod geometry;
+mod partition;
+mod surface;
+mod underlay;
+mod voids;
+mod walls;
+mod windows;
+
+#[cfg(test)]
+mod tests;
+
 use alloc::vec::Vec;
 
-use aether_capabilities::render::{DrawTriangle, Vertex};
+use aether_capabilities::render::DrawTriangle;
 
-use crate::world::{
-    CELLS_PER_CHUNK, CellPos, ChunkPos, Material, STEP_MAX_OCTIMETERS, SUBCELLS_PER_CELL_EDGE,
-    World,
-};
-use contour::{
-    GridPlacement, SmoothParams, label_case, label_window_polys, march_grid, minimize_corners,
-    push_quad, repartition,
-};
-use style::{StyleTable, flat_color, hsl_to_linear_rgb};
-
-/// Cells along one chunk edge, as a plain `i32` for loop bounds.
-const EDGE: i32 = CELLS_PER_CHUNK;
-
-/// Subcells along one cell edge, as `i32`.
-const SUB: i32 = SUBCELLS_PER_CELL_EDGE as i32;
-
-/// Subcells along one chunk edge (`16 * 16 = 256`).
-const SUBCELLS_PER_CHUNK_EDGE: i32 = EDGE * SUB;
-
-/// Octimeters per subcell (`256 / SUB = 16`).
-const OCTIMETERS_PER_SUBCELL: i32 = 256 / SUB;
-
-/// Octimeters per meter, for the octimeter-to-meter conversion at vertex
-/// emit.
-const OCTIMETERS_PER_METER: f32 = 256.0;
-
-/// How far in octimeters an unbounded-void wall drops below its top edge —
-/// the border-skirt fallback for the one void case with no far rim within
-/// the fill-over march bound (a void that reaches the world border). A
-/// bounded void joint closes instead as a real groove: wall down to the void
-/// floor, floor across, wall back up (see [`emit_void_floors`]). The skirt
-/// reads as thick ground rather than a paper-thin lip.
-const WALL_VOID_SKIRT_OCTIMETERS: i32 = 512;
-
-/// Upsample factor for the partition contour grid. At `SUB = 16`, the base
-/// subcell lattice is already finer than the old upsampled grid, and keeping
-/// the extra pass makes dense remesh tests miss CI time budgets.
-const CONTOUR_UPSAMPLE: usize = 1;
-
-/// Apron cap in subcells (two cells) so a chunk's reads stay within the
-/// eight-neighbor remesh the `R = 1` invalidation covers.
-const MAX_APRON_SUBCELLS: i32 = 2 * SUB;
-
-/// The coverage layer lift: two octimeters over the underlay datum, matching
-/// the retired overlay rim lift without restoring the old rim/body stack.
-const COVERAGE_LIFT: f32 = 2.0 / OCTIMETERS_PER_METER;
+use crate::world::{ChunkPos, World};
+use coverage::mesh_coverage;
+use style::StyleTable;
+use underlay::mesh_underlay;
+use walls::emit_walls;
 
 /// Mesh one chunk into its flat-color base triangle list. Pure — no wgpu,
 /// no ctx — so it is unit-testable host-side. Reads neighbor cells through
@@ -134,3226 +105,4 @@ pub fn mesh_chunk(world: &World, at: ChunkPos, styles: &StyleTable) -> Vec<DrawT
     mesh_coverage(world, at, styles, &mut tris);
     emit_walls(world, at, styles, partition.as_ref(), &mut tris);
     tris
-}
-
-/// The repartitioned material grid the underlay pass marches its caps from,
-/// handed to the wall pass so it lofts its curtains from the same samples —
-/// the shared grid is what makes a wall top land exactly on a cap contour
-/// vertex. Carries the display labels, the grid width, and the octimeter
-/// placement (apron offset and sample step) the boundary walk reconstructs
-/// window positions from.
-struct DisplayPartition {
-    display: Vec<u8>,
-    gw: usize,
-    apron: i32,
-    step_oct: i32,
-}
-
-/// One cell's bilinear surface patch: the four plate-resolved corner
-/// heights ([`World::cell_corner_heights`]) plus the height and shade
-/// evaluations the mesher builds vertices from. Owner-pinned — a vertex
-/// on a cliff edge evaluated through the high cell's patch reads the
-/// high plate, which is exactly how the drawn break lands on the cliff
-/// line.
-struct CellLift {
-    x0: f32,
-    z0: f32,
-    corners: [f32; 4],
-}
-
-impl CellLift {
-    fn of(world: &World, cell: CellPos) -> Self {
-        Self {
-            x0: cell.x as f32,
-            z0: cell.z as f32,
-            corners: world.cell_corner_heights(cell),
-        }
-    }
-
-    /// The patch height at `(wx, wz)` meters, coordinates clamped to the
-    /// cell span — [`World::surface_height_in`] over the cached corners.
-    fn y(&self, wx: f32, wz: f32) -> f32 {
-        let fx = (wx - self.x0).clamp(0.0, 1.0);
-        let fz = (wz - self.z0).clamp(0.0, 1.0);
-        let bottom = self.corners[0] + (self.corners[1] - self.corners[0]) * fx;
-        let top = self.corners[2] + (self.corners[3] - self.corners[2]) * fx;
-        bottom + (top - bottom) * fz
-    }
-}
-
-/// One subcell's bilinear surface patch — the point-lattice analogue of
-/// [`CellLift`], spanning `1 / SUB` m. Built only where a cell carries
-/// authored height relief; its four corners are the point-plate heights
-/// ([`World::subcell_corner_heights`]) the mesher lifts per-point caps and
-/// wall tops through. [`World::surface_height_in`]'s relief branch resolves
-/// the identical patch, so drawn and stood-on agree over authored relief.
-struct SubPatch {
-    x0: f32,
-    z0: f32,
-    corners: [f32; 4],
-}
-
-impl SubPatch {
-    fn of(world: &World, cell: CellPos, sub_x: i32, sub_z: i32) -> Self {
-        Self {
-            x0: cell.x as f32 + sub_x as f32 / SUB as f32,
-            z0: cell.z as f32 + sub_z as f32 / SUB as f32,
-            corners: world.subcell_corner_heights(cell, sub_x, sub_z),
-        }
-    }
-
-    /// The subcell of `cell` containing `(wx, wz)`, coordinates clamped to
-    /// the cell span so an off-cell caller reads the nearest edge subcell —
-    /// the same selection [`World::surface_height_in`] makes.
-    fn containing(world: &World, cell: CellPos, wx: f32, wz: f32) -> Self {
-        let sub = SUB as f32;
-        let local_x = ((wx - cell.x as f32) * sub).clamp(0.0, sub);
-        let local_z = ((wz - cell.z as f32) * sub).clamp(0.0, sub);
-        let sub_x = floor_to_i32(local_x).clamp(0, SUB - 1);
-        let sub_z = floor_to_i32(local_z).clamp(0, SUB - 1);
-        Self::of(world, cell, sub_x, sub_z)
-    }
-
-    fn y(&self, wx: f32, wz: f32) -> f32 {
-        let sub = SUB as f32;
-        let fx = ((wx - self.x0) * sub).clamp(0.0, 1.0);
-        let fz = ((wz - self.z0) * sub).clamp(0.0, 1.0);
-        let bottom = self.corners[0] + (self.corners[1] - self.corners[0]) * fx;
-        let top = self.corners[2] + (self.corners[3] - self.corners[2]) * fx;
-        bottom + (top - bottom) * fz
-    }
-}
-
-/// The lift for a vertex of label geometry owned by `owner`: position-pure
-/// through the vertex's own (floor) cell, unless that cell stands a cliff
-/// apart from the owner — then the owner's clamped patch wins, so a
-/// boundary polygon at a cliff stays on its own side of the break instead
-/// of stretching down the face (the wall draws the face). On continuous
-/// ground the two forms agree and the rule is invisible. A material-break
-/// crossing does not come here — it carries its side as data and lifts
-/// through [`anchored_lift`].
-fn label_lift(world: &World, owner: CellPos, wx: f32, wz: f32) -> f32 {
-    let cell = CellPos {
-        x: floor_to_i32(wx),
-        z: floor_to_i32(wz),
-    };
-    if cell != owner && world.edge_is_cliff(cell, owner) {
-        if world.cell_has_height_relief(owner) {
-            return SubPatch::containing(world, owner, wx, wz).y(wx, wz);
-        }
-        return CellLift::of(world, owner).y(wx, wz);
-    }
-    if world.cell_has_height_relief(cell) {
-        return SubPatch::containing(world, cell, wx, wz).y(wx, wz);
-    }
-    CellLift::of(world, cell).y(wx, wz)
-}
-
-/// The lift for a material-break crossing whose side is known as **data**:
-/// `anchor_oct` is the display-grid sample (a window corner) on the
-/// vertex's own side of the crossed edge, and over authored relief the
-/// vertex lifts through that sample's subcell patch — so the high flap
-/// holds its plate, the low flap holds the ground, and the marched wall
-/// closes the gap on the same anchors. The side is never inferred from the
-/// vertex position: a smoothing-displaced crossing sits off the subcell
-/// lattice, where positional inference reads whichever subcell the vertex
-/// floors into and collapses both flaps onto one plate. A crossing lies
-/// within half a sample of its anchor, so the anchored patch never
-/// extrapolates. The whole-cell lift resolves through the anchor sample's
-/// cell too — so a marched wall top or base on a material boundary with no
-/// subcell relief lands on the anchored side's committed plate, rather than
-/// collapsing onto the window-center `owner` wherever that owner falls on
-/// the low side of the boundary (the bug that left relief-free
-/// material-boundary cliffs with see-through faces). The relief owner-clamp
-/// branch is unchanged.
-fn anchored_lift(world: &World, owner: CellPos, anchor_oct: [i32; 2], wx: f32, wz: f32) -> f32 {
-    let cell = CellPos {
-        x: floor_to_i32(wx),
-        z: floor_to_i32(wz),
-    };
-    // The cell owning the anchor sample — the vertex's own side of the
-    // crossed edge. The whole-cell lift resolves through it (not the
-    // window-center `owner`, nor the on-boundary midpoint `cell`), so a
-    // marched wall top/base lands on the anchored side's plate even with no
-    // subcell relief. A window-center-owned lift collapses the wall wherever
-    // the owner falls on the low side of the boundary.
-    let anchor_cell = CellPos {
-        x: anchor_oct[0].div_euclid(256),
-        z: anchor_oct[1].div_euclid(256),
-    };
-    if cell != owner && world.edge_is_cliff(cell, owner) {
-        if world.cell_has_height_relief(owner) {
-            return SubPatch::containing(world, owner, wx, wz).y(wx, wz);
-        }
-        return CellLift::of(world, anchor_cell).y(wx, wz);
-    }
-    if world.cell_has_height_relief(cell) {
-        let sub_x = anchor_oct[0].rem_euclid(256) / OCTIMETERS_PER_SUBCELL;
-        let sub_z = anchor_oct[1].rem_euclid(256) / OCTIMETERS_PER_SUBCELL;
-        return SubPatch::of(world, anchor_cell, sub_x, sub_z).y(wx, wz);
-    }
-    CellLift::of(world, anchor_cell).y(wx, wz)
-}
-
-/// The break lines crossing a window's interior, per axis, in octimeters:
-/// `Some(midline)` when the window's corner-sample point levels split past
-/// the step ceiling across that axis. Levels differ across an axis only
-/// when the two sample columns (rows) sit in different subcells, so a
-/// returned midline is always a subcell lattice line — the line the break
-/// stands on and the relief walls loft from. A flap polygon spanning a
-/// returned line must be clipped there ([`emit_clipped_flap`]): a cap fan
-/// must never connect vertices whose plates split.
-///
-/// Reads the point-surface levels directly rather than gating on
-/// [`World::cell_has_height_relief`], so a **whole-cell** height cliff — a
-/// plate break at a cell boundary with no subcell relief — also returns a
-/// break line. That is what lets a chamfered relief-free material boundary
-/// clip its overhang flap so the sliver closer ([`emit_clip_gap_walls`]) can
-/// seal it; a flat cell splits nothing (equal levels) and stays break-free.
-fn window_break_lines(
-    world: &World,
-    x_lo: i32,
-    z_lo: i32,
-    step_oct: i32,
-) -> (Option<i32>, Option<i32>) {
-    let x_hi = x_lo + step_oct;
-    let z_hi = z_lo + step_oct;
-    // Corner-sample point levels in [BL, BR, TR, TL] order.
-    let level = [[x_lo, z_lo], [x_hi, z_lo], [x_hi, z_hi], [x_lo, z_hi]]
-        .map(|[px, pz]| point_surface_level_at(world, px, pz));
-    let x_break = (level[0] - level[1]).abs() > STEP_MAX_OCTIMETERS
-        || (level[3] - level[2]).abs() > STEP_MAX_OCTIMETERS;
-    let z_break = (level[0] - level[3]).abs() > STEP_MAX_OCTIMETERS
-        || (level[1] - level[2]).abs() > STEP_MAX_OCTIMETERS;
-    let half = step_oct / 2;
-    (
-        x_break.then_some(x_lo + half),
-        z_break.then_some(z_lo + half),
-    )
-}
-
-/// Split a convex polygon (octimeter coordinates) by the axis-aligned line
-/// `coord[axis] = line`, returning the `(below, above)` sides in boundary
-/// order. A vertex exactly on the line joins both sides, and a crossed
-/// edge gains its intersection on both — so the two fragments share their
-/// cut edge vertex-for-vertex.
-fn split_poly_at(poly: &[[f32; 2]], axis: usize, line: f32) -> (Vec<[f32; 2]>, Vec<[f32; 2]>) {
-    let mut below = Vec::new();
-    let mut above = Vec::new();
-    for (i, &a) in poly.iter().enumerate() {
-        let b = poly[(i + 1) % poly.len()];
-        let da = a[axis] - line;
-        let db = b[axis] - line;
-        if da <= 0.0 {
-            below.push(a);
-        }
-        if da >= 0.0 {
-            above.push(a);
-        }
-        if (da < 0.0 && db > 0.0) || (da > 0.0 && db < 0.0) {
-            let t = da / (da - db);
-            let mut p = [0.0f32; 2];
-            p[axis] = line;
-            p[1 - axis] = a[1 - axis] + (b[1 - axis] - a[1 - axis]) * t;
-            below.push(p);
-            above.push(p);
-        }
-    }
-    (below, above)
-}
-
-/// Twice the area of a polygon (shoelace, absolute) — the degenerate-
-/// fragment filter for [`emit_clipped_flap`].
-fn poly_area_doubled(poly: &[[f32; 2]]) -> f32 {
-    let mut sum = 0.0f32;
-    for (i, &a) in poly.iter().enumerate() {
-        let b = poly[(i + 1) % poly.len()];
-        sum += a[0] * b[1] - b[0] * a[1];
-    }
-    sum.abs()
-}
-
-/// One clipped flap fragment: its polygon (octimeter coordinates) and its
-/// side of each clipped break line per axis (`false` = below) — the datum
-/// [`fragment_lift`] resolves an on-line vertex's subcell through.
-type FlapFragment = (Vec<[f32; 2]>, [Option<(i32, bool)>; 2]);
-
-/// Clip one label flap polygon (octimeter coordinates) along the window's
-/// break midlines, so no fan triangle can connect vertices whose plates
-/// split — the slanted bridge a whole-window fan draws across an internal
-/// break. With no break lines the polygon passes through whole; degenerate
-/// slivers of a clip are dropped.
-fn clip_window_poly(
-    poly: &[[f32; 2]],
-    clip_x: Option<i32>,
-    clip_z: Option<i32>,
-) -> Vec<FlapFragment> {
-    let mut fragments: Vec<FlapFragment> = vec![(poly.to_vec(), [None, None])];
-    for (axis, clip) in [clip_x, clip_z].into_iter().enumerate() {
-        let Some(line) = clip else {
-            continue;
-        };
-        fragments = fragments
-            .into_iter()
-            .flat_map(|(frag, sides)| {
-                let (below, above) = split_poly_at(&frag, axis, line as f32);
-                let mut lo_sides = sides;
-                lo_sides[axis] = Some((line, false));
-                let mut hi_sides = sides;
-                hi_sides[axis] = Some((line, true));
-                [(below, lo_sides), (above, hi_sides)]
-            })
-            .collect();
-    }
-    fragments.retain(|(frag, _)| frag.len() >= 3 && poly_area_doubled(frag) >= 1.0);
-    fragments
-}
-
-/// Close the vertical gaps a clipped flap opens **inside itself**: where
-/// two fragments of one flap meet across a clipped break line at split
-/// plates, no other wall class stands — the display is uniform there (one
-/// flap, so no marched boundary), and where the flanking authored
-/// materials differ the relief walk skips the edge as marched territory.
-/// The regime is the marching chamfer at a silhouette corner: the cut
-/// leaves a sliver of this label's display over the far side's raised (or
-/// dropped) fabric, honestly lifted to that fabric's plate, and its
-/// lattice-line edges would otherwise be open cracks. The shared chord of
-/// each adjacent fragment pair is walled from the higher side's lift down to
-/// the lower's committed edge — gated to authored-material-differing lines so
-/// a same-material break stays the closure walk's (and only its) face.
-#[allow(clippy::too_many_lines)] // one pass: pair fragments, gate, loft
-fn emit_clip_gap_walls(
-    world: &World,
-    owner: CellPos,
-    fragments: &[FlapFragment],
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    let mut color: Option<[f32; 3]> = None;
-    for axis in 0..2 {
-        for (fa, sa) in fragments {
-            let Some((line, false)) = sa[axis] else {
-                continue;
-            };
-            for (fb, sb) in fragments {
-                if sb[axis] != Some((line, true)) || sb[1 - axis] != sa[1 - axis] {
-                    continue;
-                }
-                // The shared chord: each side's on-line vertex extent along
-                // the other axis, overlapped.
-                let extent = |frag: &[[f32; 2]]| -> Option<(f32, f32)> {
-                    let mut lo = f32::MAX;
-                    let mut hi = f32::MIN;
-                    for v in frag {
-                        if (v[axis] - line as f32).abs() < 0.5 {
-                            lo = lo.min(v[1 - axis]);
-                            hi = hi.max(v[1 - axis]);
-                        }
-                    }
-                    (hi > lo).then_some((lo, hi))
-                };
-                let (Some((a_lo, a_hi)), Some((b_lo, b_hi))) = (extent(fa), extent(fb)) else {
-                    continue;
-                };
-                let lo = a_lo.max(b_lo);
-                let hi = a_hi.min(b_hi);
-                if hi - lo < 0.5 {
-                    continue;
-                }
-                // Authored materials across the line at the chord: a
-                // same-material break is the relief walk's face, not ours.
-                let lattice = line / OCTIMETERS_PER_SUBCELL;
-                let sub_other = floor_to_i32((lo + hi) * 0.5 / OCTIMETERS_PER_SUBCELL as f32);
-                let material_of = |sub_axis: i32| {
-                    let (sx, sz) = if axis == 0 {
-                        (sub_axis, sub_other)
-                    } else {
-                        (sub_other, sub_axis)
-                    };
-                    world.underlay_point(
-                        CellPos {
-                            x: sx.div_euclid(SUB),
-                            z: sz.div_euclid(SUB),
-                        },
-                        sx.rem_euclid(SUB),
-                        sz.rem_euclid(SUB),
-                    )
-                };
-                if material_of(lattice - 1) == material_of(lattice) {
-                    continue;
-                }
-                // Lift the chord endpoints through both fragments' sides.
-                let endpoint = |other: f32| -> [f32; 2] {
-                    let mut p = [0.0f32; 2];
-                    p[axis] = line as f32;
-                    p[1 - axis] = other;
-                    p
-                };
-                let lift = |p: [f32; 2], sides: [Option<(i32, bool)>; 2]| {
-                    fragment_lift(
-                        world,
-                        owner,
-                        sides,
-                        p[0] / OCTIMETERS_PER_METER,
-                        p[1] / OCTIMETERS_PER_METER,
-                    )
-                };
-                let (c0, c1) = (endpoint(lo), endpoint(hi));
-                let below = [lift(c0, *sa), lift(c1, *sa)];
-                let above = [lift(c0, *sb), lift(c1, *sb)];
-                if (below[0] - above[0]).abs() < f32::EPSILON
-                    && (below[1] - above[1]).abs() < f32::EPSILON
-                {
-                    continue; // the plates agree — no gap to close
-                }
-                let (top, base) = if below[0] + below[1] > above[0] + above[1] {
-                    (below, above)
-                } else {
-                    (above, below)
-                };
-                let face = *color
-                    .get_or_insert_with(|| flat_color(styles.get(world.cliff_material(owner))));
-                push_wall_quad(
-                    tris,
-                    [
-                        c0[0] / OCTIMETERS_PER_METER,
-                        c0[1] / OCTIMETERS_PER_METER,
-                        top[0],
-                    ],
-                    [
-                        c1[0] / OCTIMETERS_PER_METER,
-                        c1[1] / OCTIMETERS_PER_METER,
-                        top[1],
-                    ],
-                    base[0],
-                    base[1],
-                    face,
-                );
-            }
-        }
-    }
-}
-
-/// The lift for a vertex of a clipped flap fragment: position-pure through
-/// the vertex's own (floor) subcell, except that a vertex lying exactly on
-/// a clipped break line reads the subcell on the **fragment's** side of it
-/// (`sides` per axis, from [`emit_clipped_flap`]) — the high fragment holds
-/// its plate, the low fragment holds the ground, and the wall classes close
-/// the vertical gap on the same subcells. Off the break lines the two forms
-/// agree wherever the plates connect, so continuous relief stays seamless.
-/// The whole-cell (relief-free) lift resolves through the fragment's own
-/// side cell as well, so a clipped relief-free material-boundary fragment
-/// reads its own plate; off any break line the side cell is the vertex's own
-/// floor cell, leaving an unclipped relief-free vertex unchanged.
-fn fragment_lift(
-    world: &World,
-    owner: CellPos,
-    sides: [Option<(i32, bool)>; 2],
-    wx: f32,
-    wz: f32,
-) -> f32 {
-    let cell = CellPos {
-        x: floor_to_i32(wx),
-        z: floor_to_i32(wz),
-    };
-    // The subcell on the fragment's own side of any clipped break line (per
-    // axis): on the line, `sides` says which side; off it, the floor subcell.
-    let sub_of = |w: f32, side: Option<(i32, bool)>| -> i32 {
-        if let Some((line, above)) = side {
-            let oct = w * OCTIMETERS_PER_METER;
-            if (oct - line as f32).abs() < 0.5 {
-                let lattice = line / OCTIMETERS_PER_SUBCELL;
-                return if above { lattice } else { lattice - 1 };
-            }
-        }
-        floor_to_i32(w * SUB as f32)
-    };
-    let sx = sub_of(wx, sides[0]);
-    let sz = sub_of(wz, sides[1]);
-    // The plate the fragment sits on — its own side cell resolved from the
-    // clipped break sides, not the window-center owner. This is what a
-    // relief-free clipped material-boundary fragment must read so its lift
-    // lands on the correct plate rather than collapsing both sides to one
-    // height (the twin of the `anchored_lift` bug, which left chamfered
-    // relief-free corners with unclosed overhang slivers).
-    let side_cell = CellPos {
-        x: sx.div_euclid(SUB),
-        z: sz.div_euclid(SUB),
-    };
-    if cell != owner && world.edge_is_cliff(cell, owner) {
-        if world.cell_has_height_relief(owner) {
-            return SubPatch::containing(world, owner, wx, wz).y(wx, wz);
-        }
-        return CellLift::of(world, side_cell).y(wx, wz);
-    }
-    if world.cell_has_height_relief(cell) {
-        return SubPatch::of(world, side_cell, sx.rem_euclid(SUB), sz.rem_euclid(SUB)).y(wx, wz);
-    }
-    CellLift::of(world, side_cell).y(wx, wz)
-}
-
-/// Floor to `i32` — `as i32` truncates toward zero, which is wrong for
-/// negative world coordinates, so step down when it rounded up.
-fn floor_to_i32(v: f32) -> i32 {
-    let t = v as i32;
-    if (t as f32) > v { t - 1 } else { t }
-}
-
-/// The effective point surface level in octimeters at octimeter position
-/// `(px, pz)` — the cell and subcell it floors into, resolved through
-/// [`World::point_surface_level`]. The marched wall gate samples this so its
-/// break reads the same authored relief the cap drew.
-fn point_surface_level_at(world: &World, px: i32, pz: i32) -> i32 {
-    let cell = CellPos {
-        x: px.div_euclid(256),
-        z: pz.div_euclid(256),
-    };
-    let sub_x = px.rem_euclid(256) / OCTIMETERS_PER_SUBCELL;
-    let sub_z = pz.rem_euclid(256) / OCTIMETERS_PER_SUBCELL;
-    world.point_surface_level(cell, sub_x, sub_z)
-}
-
-/// The per-side pooled-rim strengths `[left, right, top, bottom]` for an
-/// interior cell of `material`: a proportional rim where a same-material
-/// neighbor's hue steps past the blob-merge threshold. Material-change
-/// edges rim through the partition's marched band, never here. A pure
-/// The partition's inputs at subcell expression: the cascade-resolved
-/// material id for every sample of the chunk plus its apron, the
-/// smoothing-disabled params (zero iterations — the base render is crisp,
-/// no chamfer), and the frozen mask — a sample flanking a point-height
-/// break freezes so any future smoothing pass could never move the paint
-/// boundary across a physical cliff ([`repartition`]'s barrier). `None`
-/// when the whole area is Void — nothing to mesh.
-fn partition_inputs(
-    world: &World,
-    at: ChunkPos,
-    apron: i32,
-    n: usize,
-) -> Option<(Vec<u8>, Vec<SmoothParams>, Vec<bool>)> {
-    let mut ids = vec![0u8; n * n];
-    // The base render disables smoothing: every sample carries zero
-    // iterations, so `repartition` runs its crisp path (no chamfer). The
-    // per-cell smoothing plane and material smoothing defaults are ignored
-    // here; the smoothing rung reintroduces them.
-    let params = vec![
-        SmoothParams {
-            iterations: 0,
-            smoothing_degrees: 90,
-        };
-        n * n
-    ];
-    let mut frozen = vec![false; n * n];
-    let mut any = false;
-    for sj in -apron..SUBCELLS_PER_CHUNK_EDGE + apron {
-        for si in -apron..SUBCELLS_PER_CHUNK_EDGE + apron {
-            let cell = CellPos {
-                x: at.x * EDGE + si.div_euclid(SUB),
-                z: at.z * EDGE + sj.div_euclid(SUB),
-            };
-            // Sample the cell's authored point, not the whole-cell material:
-            // an all-inherit point folds back to `World::underlay(cell)`, so
-            // an unshaped world is unchanged, while an authored point moves
-            // the silhouette below cell scale.
-            let material = world.underlay_point(cell, si.rem_euclid(SUB), sj.rem_euclid(SUB));
-            let idx = (sj + apron) as usize * n + (si + apron) as usize;
-            ids[idx] = material.to_u8();
-            any |= material != Material::Void;
-            // Freeze the sample when its subcell flanks a point-height
-            // break: paint smoothing must never cross a physical cliff, or
-            // the silhouette's flank would alternate between painted steep
-            // caps and wall wedges. Pure world reads (position-based), so
-            // neighboring chunks agree over the shared apron. Reads the point
-            // levels directly rather than gating on relief, so a **whole-cell**
-            // cliff freezes the contour too: without this the chamfer smooths
-            // across a relief-free material cliff, flipping the corner to the
-            // low neighbor's paint over the high plate (the floating sliver)
-            // and leaving its flank unwalled. A flat neighborhood reads equal
-            // levels and stays unfrozen, so an unshaped world is unchanged.
-            let gx = at.x * EDGE * SUB + si;
-            let gz = at.z * EDGE * SUB + sj;
-            let own_level = point_surface_level_at(
-                world,
-                gx * OCTIMETERS_PER_SUBCELL,
-                gz * OCTIMETERS_PER_SUBCELL,
-            );
-            frozen[idx] =
-                [(1, 0), (-1, 0), (0, 1), (0, -1)]
-                    .into_iter()
-                    .any(|(dx, dz): (i32, i32)| {
-                        let level = point_surface_level_at(
-                            world,
-                            (gx + dx) * OCTIMETERS_PER_SUBCELL,
-                            (gz + dz) * OCTIMETERS_PER_SUBCELL,
-                        );
-                        (own_level - level).abs() > STEP_MAX_OCTIMETERS
-                    });
-        }
-    }
-    any.then_some((ids, params, frozen))
-}
-
-/// Emit the underlay as a partition of one flat surface: repartition the
-/// cascade-resolved material grid (smoothing off — crisp marching-squares
-/// contours, no chamfer), then tile the ground exactly — flat keyed quilt
-/// cells wherever a cell and its one-sample surround are uniformly one
-/// material, and per-window marching polygons everywhere else, all at
-/// `y = 0` with no lifts. Every decision is a pure function of world
-/// coordinates, so two chunks emit identical geometry over their shared
-/// apron and the overlap is invisible. Returns the display grid (the plain
-/// material grid) so the wall pass lofts from the same samples; `None` when
-/// the whole chunk plus apron is Void (nothing to mesh).
-#[allow(clippy::too_many_lines)] // one underlay pass: partition, tile interiors, march windows
-fn mesh_underlay(
-    world: &World,
-    at: ChunkPos,
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) -> Option<DisplayPartition> {
-    let apron = MAX_APRON_SUBCELLS;
-    let n = (SUBCELLS_PER_CHUNK_EDGE + 2 * apron) as usize;
-    let (ids, params, frozen) = partition_inputs(world, at, apron, n)?;
-
-    let upsample = CONTOUR_UPSAMPLE;
-    let (grid, gw, _gh) = repartition(&ids, n, n, upsample, &params, &frozen);
-    let u = upsample as i32;
-    let step_oct = OCTIMETERS_PER_SUBCELL / u;
-
-    // The base render has no rim/body split, so the display grid is the
-    // repartitioned material grid directly — each material is one flat label.
-    let display = grid;
-
-    // Interior cells: the cell's sample block plus a one-sample surround
-    // is uniformly its own material. Classified for every cell whose
-    // surround fits the grid (local -1..EDGE), identically on both sides of
-    // a chunk border, so the window skip below never disagrees with a
-    // neighbor's cell quad.
-    let lo = -1i32;
-    let hi = EDGE;
-    let cells_w = (hi - lo) as usize;
-    let mut interior = vec![false; cells_w * cells_w];
-    for lz in lo..hi {
-        for lx in lo..hi {
-            let cell = CellPos {
-                x: at.x * EDGE + lx,
-                z: at.z * EDGE + lz,
-            };
-            let m = world.underlay(cell).to_u8();
-            if m == 0 {
-                continue;
-            }
-            let x0 = (lx * SUB + apron) * u;
-            let z0 = (lz * SUB + apron) * u;
-            let span = SUB * u;
-            let uniform = ((z0 - 1)..=(z0 + span)).all(|gz| {
-                ((x0 - 1)..=(x0 + span)).all(|gx| {
-                    gx >= 0
-                        && gz >= 0
-                        && (gx as usize) < gw
-                        && (gz as usize) < gw
-                        && display[gz as usize * gw + gx as usize] == m
-                })
-            });
-            interior[(lz - lo) as usize * cells_w + (lx - lo) as usize] = uniform;
-        }
-    }
-
-    // Chunk-local interior cells emit the flat keyed quilt cell.
-    for lz in 0..EDGE {
-        for lx in 0..EDGE {
-            if !interior[(lz - lo) as usize * cells_w + (lx - lo) as usize] {
-                continue;
-            }
-            let cell = CellPos {
-                x: at.x * EDGE + lx,
-                z: at.z * EDGE + lz,
-            };
-            let material = world.underlay(cell);
-            // Authored relief tessellates the cap to subcell resolution so
-            // its per-point heights and breaks show; a flat cell keeps the
-            // whole-cell fast path (byte-identical to a world with no relief).
-            if world.cell_has_height_relief(cell) {
-                emit_underlay_cell_subdivided(world, material, cell, styles, tris);
-                continue;
-            }
-            let lift = CellLift::of(world, cell);
-            emit_underlay_cell(material, cell.x, cell.z, &lift, styles, tris);
-        }
-    }
-
-    emit_partition_windows(
-        world, at, &display, gw, apron, step_oct, &interior, lo, styles, tris,
-    );
-
-    // The fill-over floor caps for enclosed Void joints — the flat groove
-    // bottoms the marched closure's Void walls drop to.
-    emit_void_floors(world, at, styles, tris);
-
-    Some(DisplayPartition {
-        display,
-        gw,
-        apron,
-        step_oct,
-    })
-}
-
-/// Where a chunk's partition grid lands in the world: sample `(0, 0)`'s
-/// octimeter center, offset back by the apron and forward half a step so
-/// each sample sits at a subcell-lattice center, at `step_oct` spacing.
-/// Both the partition-window and encroachment-flap passes place their
-/// marches through this, so they agree on the lattice sample-for-sample.
-fn chunk_placement(at: ChunkPos, apron: i32, step_oct: i32) -> GridPlacement {
-    let base_oct = [
-        at.x * SUBCELLS_PER_CHUNK_EDGE * OCTIMETERS_PER_SUBCELL,
-        at.z * SUBCELLS_PER_CHUNK_EDGE * OCTIMETERS_PER_SUBCELL,
-    ];
-    GridPlacement {
-        origin_oct: [
-            base_oct[0] - apron * OCTIMETERS_PER_SUBCELL + step_oct / 2,
-            base_oct[1] - apron * OCTIMETERS_PER_SUBCELL + step_oct / 2,
-        ],
-        step_oct,
-    }
-}
-
-/// Distinct non-void overlay materials present in the chunk, in stable
-/// material-id order.
-fn overlay_materials(world: &World, at: ChunkPos) -> Vec<Material> {
-    let Some(chunk) = world.chunk(at) else {
-        return Vec::new();
-    };
-    let mut present = [false; 6];
-    for material in &chunk.overlay {
-        if *material != Material::Void {
-            present[*material as usize] = true;
-        }
-    }
-    present
-        .iter()
-        .enumerate()
-        .filter(|(_, seen)| **seen)
-        .map(|(id, _)| Material::from_u8_or_void(id.try_into().unwrap_or(0)))
-        .collect()
-}
-
-/// Coverage of the subcell at chunk-local index `(six, siz)` for `material`.
-/// Off-material samples read uncovered, while out-of-chunk indices resolve
-/// through [`World`] so neighboring chunks supply the apron.
-fn subcell_coverage(world: &World, at: ChunkPos, six: i32, siz: i32, material: Material) -> u8 {
-    let cell = CellPos {
-        x: at.x * EDGE + six.div_euclid(SUB),
-        z: at.z * EDGE + siz.div_euclid(SUB),
-    };
-    if world.overlay(cell) != material {
-        return 0;
-    }
-    world.overlay_coverage(cell, six, siz)
-}
-
-/// Emit a flat scalar-coverage pass for authored overlays. Each distinct
-/// overlay material gets its own scalar mask, marched through the contour
-/// library so authored `1..254` samples keep interpolated crossings.
-fn mesh_coverage(world: &World, at: ChunkPos, styles: &StyleTable, tris: &mut Vec<DrawTriangle>) {
-    let apron = MAX_APRON_SUBCELLS;
-    let n = (SUBCELLS_PER_CHUNK_EDGE + 2 * apron) as usize;
-    let params = vec![
-        SmoothParams {
-            iterations: 0,
-            smoothing_degrees: 90,
-        };
-        n * n
-    ];
-    let upsample = CONTOUR_UPSAMPLE;
-    let step_oct = OCTIMETERS_PER_SUBCELL / upsample as i32;
-    let placement = chunk_placement(at, apron, step_oct);
-    for material in overlay_materials(world, at) {
-        let mut field = vec![0; n * n];
-        for sj in -apron..SUBCELLS_PER_CHUNK_EDGE + apron {
-            for si in -apron..SUBCELLS_PER_CHUNK_EDGE + apron {
-                let idx = (sj + apron) as usize * n + (si + apron) as usize;
-                field[idx] = subcell_coverage(world, at, si, sj, material);
-            }
-        }
-        let (samples, width, height) = minimize_corners(&field, n, n, upsample, &params);
-        let style = styles.get(material);
-        let color = hsl_to_linear_rgb(style.base_hue, style.base_sat, style.base_light);
-        let vertex = |wx: f32, wz: f32| Vertex {
-            x: wx,
-            y: COVERAGE_LIFT,
-            z: wz,
-            r: color[0],
-            g: color[1],
-            b: color[2],
-        };
-        march_grid(&samples, width, height, &placement, &vertex, tris);
-    }
-}
-
-/// Emit the marching-squares windows of the partition's boundary zone.
-/// Each window is owned by exactly one chunk — the one holding the cell
-/// under its center — so the boundary zone is emitted once fleet-wide,
-/// with no cross-chunk duplicates against the fixed per-frame vertex
-/// budget. Windows fully covered by interior cell quads are skipped;
-/// uniform single-label windows coalesce into row strips (per owning
-/// cell, so the keyed color stays flat per cell); mixed windows emit each
-/// label's case polygon, saddles resolved by label order so every window
-/// tiles exactly.
-#[allow(clippy::too_many_arguments)] // one call site; the partition state travels together
-fn emit_partition_windows(
-    world: &World,
-    at: ChunkPos,
-    display: &[u8],
-    gw: usize,
-    apron: i32,
-    step_oct: i32,
-    interior: &[bool],
-    lo: i32,
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    let hi = EDGE;
-    let cells_w = (hi - lo) as usize;
-    let place = chunk_placement(at, apron, step_oct);
-    let origin_oct = place.origin_oct;
-    // A pending run of uniform windows: (label, owner cell, start wi).
-    let mut run: Option<(u8, CellPos, usize)> = None;
-    let flush = |run: &mut Option<(u8, CellPos, usize)>,
-                 end_wi: usize,
-                 wj: usize,
-                 tris: &mut Vec<DrawTriangle>| {
-        let Some((label, owner, start_wi)) = run.take() else {
-            return;
-        };
-        let rect = [
-            origin_oct[0] + start_wi as i32 * step_oct,
-            origin_oct[1] + wj as i32 * step_oct,
-            origin_oct[0] + end_wi as i32 * step_oct,
-            origin_oct[1] + (wj + 1) as i32 * step_oct,
-        ];
-        emit_label_quad(world, label, owner, rect, styles, tris);
-    };
-    let windows = gw - 1;
-    for wj in 0..windows {
-        for wi in 0..windows {
-            let x_lo = origin_oct[0] + wi as i32 * step_oct;
-            let z_lo = origin_oct[1] + wj as i32 * step_oct;
-            // Ownership: the cell under the window's center. Emitting only
-            // chunk-local owners covers every window exactly once across
-            // the fleet, and a local owner keeps every overlapping cell
-            // within the classifiable range.
-            let x_center_cell = (x_lo + step_oct / 2).div_euclid(256);
-            let z_center_cell = (z_lo + step_oct / 2).div_euclid(256);
-            let x_owner_local = x_center_cell - at.x * EDGE;
-            let z_owner_local = z_center_cell - at.z * EDGE;
-            if !(0..EDGE).contains(&x_owner_local) || !(0..EDGE).contains(&z_owner_local) {
-                flush(&mut run, wi, wj, tris);
-                continue;
-            }
-            // The world cells this window's square overlaps.
-            let cx0 = x_lo.div_euclid(256);
-            let cx1 = (x_lo + step_oct - 1).div_euclid(256);
-            let cz0 = z_lo.div_euclid(256);
-            let cz1 = (z_lo + step_oct - 1).div_euclid(256);
-            let mut all_interior = true;
-            for cz in cz0..=cz1 {
-                for cx in cx0..=cx1 {
-                    let llx = cx - at.x * EDGE;
-                    let llz = cz - at.z * EDGE;
-                    debug_assert!(llx >= lo && llx < hi && llz >= lo && llz < hi);
-                    if !interior[(llz - lo) as usize * cells_w + (llx - lo) as usize] {
-                        all_interior = false;
-                    }
-                }
-            }
-            if all_interior {
-                // The cell quads already tile it.
-                flush(&mut run, wi, wj, tris);
-                continue;
-            }
-            let corners = [
-                display[wj * gw + wi],
-                display[wj * gw + wi + 1],
-                display[(wj + 1) * gw + wi],
-                display[(wj + 1) * gw + wi + 1],
-            ];
-            let owner = CellPos {
-                x: x_center_cell,
-                z: z_center_cell,
-            };
-            // A uniform window joins (or starts) a strip run. A run's quad
-            // takes position-pure corner heights, and its edges lie on
-            // sample-lattice lines where the bilinear surface is linear —
-            // so the merge stays crack-free on sloped ground too (the quad
-            // interior deviates from the surface by at most the tiny
-            // cross-term over one window row, and nothing else draws
-            // there). A uniform window straddling a point-height break is
-            // the exception: its quad would bridge the split plates, so it
-            // routes through the mixed emitter, whose break clipping splits
-            // it (the label's case is 15 — the full window polygon).
-            if corners[0] != 0 && corners.iter().all(|&c| c == corners[0]) {
-                let (bx, bz) = window_break_lines(world, x_lo, z_lo, step_oct);
-                if bx.is_none() && bz.is_none() {
-                    match run {
-                        Some((label, cell, _)) if label == corners[0] && cell == owner => {}
-                        _ => {
-                            flush(&mut run, wi, wj, tris);
-                            run = Some((corners[0], owner, wi));
-                        }
-                    }
-                    continue;
-                }
-            }
-            flush(&mut run, wi, wj, tris);
-            emit_mixed_window(
-                world, display, gw, wi, wj, corners, owner, &place, styles, tris,
-            );
-        }
-        flush(&mut run, windows, wj, tris);
-    }
-}
-
-/// Emit one material label's quad over `rect`, colored by its material's
-/// flat color and lifted position-pure — the strip-run emit. Windows
-/// overhang their owner by half a window, so each corner reads the surface
-/// through its own cell rather than the owner's clamped patch.
-fn emit_label_quad(
-    world: &World,
-    label: u8,
-    owner: CellPos,
-    rect: [i32; 4],
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    let material = Material::from_u8_or_void(label);
-    let color = flat_color(styles.get(material));
-    let surface = |wx: f32, wz: f32| label_lift(world, owner, wx, wz);
-    emit_flat_quad(rect, color, &surface, tris);
-}
-
-/// Emit every label's case polygon for one mixed boundary window, colored
-/// by its material's flat color and lifted per vertex through the vertex's
-/// own cell. A two-label saddle resolves by label order — the higher label
-/// connects its diagonal, the lower splits — so the window always tiles
-/// exactly.
-#[allow(clippy::too_many_arguments)] // one call site; the window state travels together
-fn emit_mixed_window(
-    world: &World,
-    display: &[u8],
-    gw: usize,
-    wi: usize,
-    wj: usize,
-    corners: [u8; 4],
-    owner: CellPos,
-    place: &GridPlacement,
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    let step_oct = place.step_oct;
-    let half = step_oct / 2;
-    let x_lo = place.origin_oct[0] + wi as i32 * step_oct;
-    let z_lo = place.origin_oct[1] + wj as i32 * step_oct;
-    let x_hi = x_lo + step_oct;
-    let z_hi = z_lo + step_oct;
-    // The break midlines a flap must be clipped along — a fan across a
-    // split would draw a slanted bridge over the break instead of letting
-    // the walls close it.
-    let (clip_x, clip_z) = window_break_lines(world, x_lo, z_lo, step_oct);
-    // Window point positions in octimeters, contour point numbering
-    // (corners 0..4, edge midpoints 4..8).
-    let points = [
-        [x_lo, z_lo],
-        [x_hi, z_lo],
-        [x_hi, z_hi],
-        [x_lo, z_hi],
-        [x_lo + half, z_lo],
-        [x_hi, z_lo + half],
-        [x_lo + half, z_hi],
-        [x_lo, z_lo + half],
-    ]
-    .map(|[px, pz]| [px as f32, pz as f32]);
-    for k in 0..4 {
-        let label = corners[k];
-        if label == 0 || corners[..k].contains(&label) {
-            continue;
-        }
-        let case = label_case(display, gw, wi, wj, label);
-        let connected = if case == 5 || case == 10 {
-            // The other diagonal pair: [BR, TL] for case 5, [BL, TR] for
-            // case 10.
-            let (o1, o2) = if case == 5 {
-                (corners[1], corners[2])
-            } else {
-                (corners[0], corners[3])
-            };
-            o1 != o2 || label > o1
-        } else {
-            true
-        };
-        let material = Material::from_u8_or_void(label);
-        let color = flat_color(styles.get(material));
-        let vertex = |pos: [f32; 2], sides: [Option<(i32, bool)>; 2]| {
-            let wx = pos[0] / OCTIMETERS_PER_METER;
-            let wz = pos[1] / OCTIMETERS_PER_METER;
-            let y = fragment_lift(world, owner, sides, wx, wz);
-            Vertex {
-                x: wx,
-                y,
-                z: wz,
-                r: color[0],
-                g: color[1],
-                b: color[2],
-            }
-        };
-        for poly in label_window_polys(case, connected) {
-            if poly.is_empty() {
-                continue;
-            }
-            let pts: Vec<[f32; 2]> = poly.iter().map(|&idx| points[idx as usize]).collect();
-            let fragments = clip_window_poly(&pts, clip_x, clip_z);
-            for (frag, sides) in &fragments {
-                for k in 1..frag.len() - 1 {
-                    tris.push(DrawTriangle {
-                        verts: [
-                            vertex(frag[0], *sides),
-                            vertex(frag[k], *sides),
-                            vertex(frag[k + 1], *sides),
-                        ],
-                    });
-                }
-            }
-            if clip_x.is_some() || clip_z.is_some() {
-                emit_clip_gap_walls(world, owner, &fragments, styles, tris);
-            }
-        }
-    }
-}
-
-/// Emit one flat keyed cell: a single flat-colored quad spanning the cell
-/// on its bilinear surface patch.
-fn emit_underlay_cell(
-    material: Material,
-    cx: i32,
-    cz: i32,
-    patch: &CellLift,
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    let color = flat_color(styles.get(material));
-    let surface = |wx: f32, wz: f32| patch.y(wx, wz);
-    let x0 = cx * 256;
-    let z0 = cz * 256;
-    emit_flat_quad(
-        [x0, z0, (cx + 1) * 256, (cz + 1) * 256],
-        color,
-        &surface,
-        tris,
-    );
-}
-
-/// Emit a height-relief cell's cap as `SUB × SUB` subcell quads, each lifted
-/// through its own point patch ([`SubPatch`]) so authored subcell relief —
-/// and the breaks where adjacent points cliff — shows in the cap and the
-/// drawn height matches [`World::surface_height`] at every sample. A flat
-/// cell keeps the whole-cell fast path in [`emit_underlay_cell`].
-fn emit_underlay_cell_subdivided(
-    world: &World,
-    material: Material,
-    cell: CellPos,
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    let color = flat_color(styles.get(material));
-    for sj in 0..SUB {
-        for si in 0..SUB {
-            let patch = SubPatch::of(world, cell, si, sj);
-            let surface = |wx: f32, wz: f32| patch.y(wx, wz);
-            let x0 = cell.x * 256 + si * OCTIMETERS_PER_SUBCELL;
-            let z0 = cell.z * 256 + sj * OCTIMETERS_PER_SUBCELL;
-            emit_flat_quad(
-                [
-                    x0,
-                    z0,
-                    x0 + OCTIMETERS_PER_SUBCELL,
-                    z0 + OCTIMETERS_PER_SUBCELL,
-                ],
-                color,
-                &surface,
-                tris,
-            );
-        }
-    }
-}
-
-/// Push the two triangles of one flat-colored underlay quad spanning `rect`
-/// (`[x0, z0, x1, z1]` octimeters) on the given surface evaluator
-/// (`(wx, wz)` meters to `y`). Every corner takes the same flat `color`.
-// The quad's four corners (`a`..`d`) read clearest under these conventional
-// short names.
-#[allow(clippy::many_single_char_names)]
-fn emit_flat_quad(
-    rect: [i32; 4],
-    color: [f32; 3],
-    surface: &impl Fn(f32, f32) -> f32,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    let corner = |xo: i32, zo: i32| {
-        let wx = xo as f32 / OCTIMETERS_PER_METER;
-        let wz = zo as f32 / OCTIMETERS_PER_METER;
-        Vertex {
-            x: wx,
-            y: surface(wx, wz),
-            z: wz,
-            r: color[0],
-            g: color[1],
-            b: color[2],
-        }
-    };
-    let a = corner(rect[0], rect[1]);
-    let b = corner(rect[2], rect[1]);
-    let c = corner(rect[2], rect[3]);
-    let d = corner(rect[0], rect[3]);
-    tris.push(DrawTriangle { verts: [a, b, c] });
-    tris.push(DrawTriangle { verts: [a, c, d] });
-}
-
-const WALL_SEGMENTS: [&[(u8, u8)]; 16] = [
-    &[],               // 0
-    &[(7, 4)],         // 1  BL
-    &[(4, 5)],         // 2  BR
-    &[(7, 5)],         // 3  BL BR
-    &[(5, 6)],         // 4  TR
-    &[(7, 4), (5, 6)], // 5  BL TR (saddle)
-    &[(4, 6)],         // 6  BR TR
-    &[(7, 6)],         // 7  BL BR TR
-    &[(6, 7)],         // 8  TL
-    &[(4, 6)],         // 9  BL TL
-    &[(4, 5), (6, 7)], // 10 BR TL (saddle)
-    &[(5, 6)],         // 11 BL BR TL
-    &[(7, 5)],         // 12 TR TL
-    &[(4, 5)],         // 13 BL TR TL
-    &[(4, 7)],         // 14 BR TR TL
-    &[],               // 15
-];
-
-/// Emit the chunk's vertical cliff faces by closure: a wall lofts wherever
-/// the two sides of a shared boundary edge committed different cap heights,
-/// so the split decision lives in one place — the plate lift the caps drew
-/// from — and can never drift between a predicting wall pass and the cap it
-/// should meet. The lattice/subcell closure ([`emit_lattice_closure`]) walls
-/// every same-material cell-edge and subcell-edge break; the marched closure
-/// ([`emit_contour_closure`]) walls every material or Void boundary along
-/// the display partition's contours, dropping a fillable Void low side to
-/// its groove floor. A face lofts iff the high side committed strictly above
-/// the low side there — a legal step merges the plates (the committed edges
-/// agree) and lofts nothing, a cliff splits them and lofts a face whose top
-/// and bottom vertices are the two caps' shared-edge vertices, watertight by
-/// construction.
-fn emit_walls(
-    world: &World,
-    at: ChunkPos,
-    styles: &StyleTable,
-    partition: Option<&DisplayPartition>,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    emit_lattice_closure(world, at, styles, tris);
-    if let Some(part) = partition {
-        emit_contour_closure(world, at, part, styles, tris);
-    }
-}
-
-/// The flat wall color for a cliff owned by `cell`: the cell's cliff
-/// material's flat color. A pure function of the cell, so the two sides of a
-/// shared edge agree on the color.
-fn wall_color(world: &World, cell: CellPos, styles: &StyleTable) -> [f32; 3] {
-    flat_color(styles.get(world.cliff_material(cell)))
-}
-
-/// Close the same-material cliff faces of the chunk on the cell lattice — or,
-/// where a cell carries authored relief, on its subcell lattice, the stride
-/// its cap tessellated at. For every chunk-local cell and each of its four
-/// outgoing shared edges the two sides' committed cap edges are read through
-/// the same corner plates the caps drew from
-/// ([`World::cell_corner_heights`] / [`World::subcell_corner_heights`]); a
-/// face lofts iff this (the higher) side committed above the neighbor and the
-/// committed edges differ. A merged plate — flat ground or a legal step —
-/// leaves the edges equal and lofts nothing, so the step size is never gated
-/// directly; the plate merge is the one split decision. The high side owns
-/// the face and cells iterate chunk-local, so a chunk-border cliff lofts
-/// exactly once fleet-wide. A material or Void boundary edge is skipped here
-/// — the marched closure owns those.
-#[allow(clippy::too_many_lines)] // one closure walk over the cell and subcell strides
-fn emit_lattice_closure(
-    world: &World,
-    at: ChunkPos,
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    let sub_span = 1.0 / SUB as f32;
-    for lz in 0..EDGE {
-        for lx in 0..EDGE {
-            let cell = CellPos {
-                x: at.x * EDGE + lx,
-                z: at.z * EDGE + lz,
-            };
-            // Relief cells close on the subcell lattice (the stride their
-            // caps tessellated at); every other cell closes on the cell
-            // lattice.
-            let relief = world.cell_has_height_relief(cell);
-            let face_rgb = wall_color(world, cell, styles);
-            if relief {
-                for sj in 0..SUB {
-                    for si in 0..SUB {
-                        // The point's own material, not the cell's cascade — a
-                        // stone point on a grass cell closes its own
-                        // same-material breaks, and a Void point's rim lofts as
-                        // a marched wall.
-                        let material = world.underlay_point(cell, si, sj);
-                        if material == Material::Void {
-                            continue; // a hole's rim lofts as a marched wall
-                        }
-                        let level = world.point_surface_level(cell, si, sj);
-                        let mut cached: Option<[f32; 4]> = None;
-                        for edge in &WALL_DIRECTIONS {
-                            let gx = cell.x * SUB + si + edge.offset.0;
-                            let gz = cell.z * SUB + sj + edge.offset.1;
-                            let neighbor = CellPos {
-                                x: gx.div_euclid(SUB),
-                                z: gz.div_euclid(SUB),
-                            };
-                            let nsx = gx.rem_euclid(SUB);
-                            let nsz = gz.rem_euclid(SUB);
-                            if world.underlay_point(neighbor, nsx, nsz) != material {
-                                continue; // material / Void boundaries loft as marched walls
-                            }
-                            // Ownership: only the strictly-higher point lofts,
-                            // so each face is emitted once. The step size is
-                            // not gated — a legal step merges the point plates
-                            // and the committed-edge test below leaves nothing.
-                            if level <= world.point_surface_level(neighbor, nsx, nsz) {
-                                continue;
-                            }
-                            let top = *cached
-                                .get_or_insert_with(|| world.subcell_corner_heights(cell, si, sj));
-                            let bottom = world.subcell_corner_heights(neighbor, nsx, nsz);
-                            let y_top = [top[edge.top[0]], top[edge.top[1]]];
-                            let y_low = [bottom[edge.bottom[0]], bottom[edge.bottom[1]]];
-                            if (y_top[0] - y_low[0]).abs() < f32::EPSILON
-                                && (y_top[1] - y_low[1]).abs() < f32::EPSILON
-                            {
-                                continue; // the point plates merged — no break to close
-                            }
-                            let base_x = cell.x as f32 + si as f32 * sub_span;
-                            let base_z = cell.z as f32 + sj as f32 * sub_span;
-                            let corner_pos = |k: usize| {
-                                (
-                                    base_x + if k % 2 == 1 { sub_span } else { 0.0 },
-                                    base_z + if k >= 2 { sub_span } else { 0.0 },
-                                )
-                            };
-                            let (x0, z0) = corner_pos(edge.top[0]);
-                            let (x1, z1) = corner_pos(edge.top[1]);
-                            push_wall_quad(
-                                tris,
-                                [x0, z0, y_top[0]],
-                                [x1, z1, y_top[1]],
-                                y_low[0],
-                                y_low[1],
-                                face_rgb,
-                            );
-                        }
-                    }
-                }
-            } else {
-                let material = world.underlay(cell);
-                if material == Material::Void {
-                    continue; // a Void cell's rim is the marched closure's face
-                }
-                let cell_level = world.surface_level(cell);
-                let mut cached: Option<[f32; 4]> = None;
-                for edge in &WALL_DIRECTIONS {
-                    let neighbor = CellPos {
-                        x: cell.x + edge.offset.0,
-                        z: cell.z + edge.offset.1,
-                    };
-                    if world.underlay(neighbor) != material {
-                        continue; // a material boundary lofts as a marched wall
-                    }
-                    // Ownership: only the strictly-higher cell lofts. The step
-                    // size is not gated — a legal step merges the plates and
-                    // the committed-edge test below leaves nothing to close.
-                    if cell_level <= world.surface_level(neighbor) {
-                        continue;
-                    }
-                    let top = *cached.get_or_insert_with(|| world.cell_corner_heights(cell));
-                    let bottom = world.cell_corner_heights(neighbor);
-                    let y_top = [top[edge.top[0]], top[edge.top[1]]];
-                    let y_low = [bottom[edge.bottom[0]], bottom[edge.bottom[1]]];
-                    if (y_top[0] - y_low[0]).abs() < f32::EPSILON
-                        && (y_top[1] - y_low[1]).abs() < f32::EPSILON
-                    {
-                        continue; // the plates merged — no break to close
-                    }
-                    let corner_pos = |k: usize| {
-                        (
-                            cell.x as f32 + if k % 2 == 1 { 1.0 } else { 0.0 },
-                            cell.z as f32 + if k >= 2 { 1.0 } else { 0.0 },
-                        )
-                    };
-                    let (x0, z0) = corner_pos(edge.top[0]);
-                    let (x1, z1) = corner_pos(edge.top[1]);
-                    push_wall_quad(
-                        tris,
-                        [x0, z0, y_top[0]],
-                        [x1, z1, y_top[1]],
-                        y_low[0],
-                        y_low[1],
-                        face_rgb,
-                    );
-                }
-            }
-        }
-    }
-}
-
-/// Close the marched cliff walls: over the boundary windows of the display
-/// partition, wherever a segment separates two materials — or a material
-/// from a Void hole — whose point surface levels split the caps, drop the
-/// segment's marched contour from the high cap's committed edge to the low
-/// side's. Each window is owned by the cell under its center, the same
-/// ownership the cap walk uses; the owner is the fleet-wide dedup key (a
-/// window emits only from the chunk holding its owner) and the color source,
-/// never the gate — the split is per-segment and side-aware, reading the
-/// `a`-side corner's point level against the crossed side's, because along a
-/// contour the window center alternates sides subcell to subcell and an
-/// owner-level test would picket-fence the wall. A wall's top vertices lift
-/// through the identical anchored patch ([`anchored_lift`] on the `a`-side
-/// corner sample) the high cap's crossing vertices took — landing the top
-/// edge exactly on the cap's contour, watertight by construction under any
-/// smoothing displacement. A grounded low side takes per-endpoint bases on
-/// the low side's own committed lift over relief (a flat min base where no
-/// relief engages), coincident with the low cap. A Void low side closes to
-/// its fill-over floor ([`World::point_height`], the total height plane) when
-/// the joint is enclosed ([`void_fill_border`]) — wall down, floor across
-/// ([`emit_void_floors`]), wall up at the far rim — else drops the
-/// unbounded-void skirt [`WALL_VOID_SKIRT_OCTIMETERS`] so an open hole reads
-/// as thick ground.
-#[allow(clippy::too_many_lines)] // one boundary walk: classify sides, extract segments, close
-#[allow(clippy::similar_names)] // the segment's two endpoints read clearest as `_p` / `_q` pairs
-fn emit_contour_closure(
-    world: &World,
-    at: ChunkPos,
-    part: &DisplayPartition,
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    let display = &part.display;
-    let gw = part.gw;
-    let apron = part.apron;
-    let step_oct = part.step_oct;
-    let base_oct = [
-        at.x * SUBCELLS_PER_CHUNK_EDGE * OCTIMETERS_PER_SUBCELL,
-        at.z * SUBCELLS_PER_CHUNK_EDGE * OCTIMETERS_PER_SUBCELL,
-    ];
-    let origin_oct = [
-        base_oct[0] - apron * OCTIMETERS_PER_SUBCELL + step_oct / 2,
-        base_oct[1] - apron * OCTIMETERS_PER_SUBCELL + step_oct / 2,
-    ];
-    let half = step_oct / 2;
-    let windows = gw - 1;
-    for wj in 0..windows {
-        for wi in 0..windows {
-            let x_lo = origin_oct[0] + wi as i32 * step_oct;
-            let z_lo = origin_oct[1] + wj as i32 * step_oct;
-            let x_hi = x_lo + step_oct;
-            let z_hi = z_lo + step_oct;
-            // Ownership: the cell under the window center, chunk-local only —
-            // covers every window exactly once across the fleet and matches
-            // the cap walk's owner, so the lofted top reuses the cap's patch.
-            let owner = CellPos {
-                x: (x_lo + half).div_euclid(256),
-                z: (z_lo + half).div_euclid(256),
-            };
-            if !(0..EDGE).contains(&(owner.x - at.x * EDGE))
-                || !(0..EDGE).contains(&(owner.z - at.z * EDGE))
-            {
-                continue;
-            }
-            // Corner materials in order [BL, BR, TR, TL] (the case-bit
-            // order), folding the rim/body display split back to the plain
-            // material — a wall follows a material boundary, not a rim edge.
-            let mats = [
-                display[wj * gw + wi].div_ceil(2),
-                display[wj * gw + wi + 1].div_ceil(2),
-                display[(wj + 1) * gw + wi + 1].div_ceil(2),
-                display[(wj + 1) * gw + wi].div_ceil(2),
-            ];
-            if mats.iter().all(|&m| m == mats[0]) {
-                continue; // uniform window — no material boundary to loft
-            }
-            // Does authored relief engage anywhere this window can read? The
-            // per-endpoint wall bases below switch on this, so relief-free
-            // worlds keep their flat base byte-identically.
-            let relief = world.cell_has_height_relief(owner);
-            // Corner sample positions and point surface levels, same order
-            // as `mats`. Point levels (not the cell level) so an authored
-            // break inside a cell — a silhouette raised by deltas alone,
-            // the cell height flat — lofts its wall, and the gate reads the
-            // sample the cap actually drew. The positions double as the
-            // wall's side anchors ([`anchored_lift`]).
-            let corner_oct = [[x_lo, z_lo], [x_hi, z_lo], [x_hi, z_hi], [x_lo, z_hi]];
-            let point_level = corner_oct.map(|[cx, cz]| point_surface_level_at(world, cx, cz));
-            let mid = |m: u8| -> [i32; 2] {
-                match m {
-                    4 => [x_lo + half, z_lo],
-                    5 => [x_hi, z_lo + half],
-                    6 => [x_lo + half, z_hi],
-                    _ => [x_lo, z_lo + half],
-                }
-            };
-            let edge_corners = |m: u8| -> (usize, usize) {
-                match m {
-                    4 => (0, 1),
-                    5 => (1, 2),
-                    6 => (2, 3),
-                    _ => (3, 0),
-                }
-            };
-            // Wall color: the owning cell's cliff material as a flat color —
-            // the owner is the color source even when the segment's high side
-            // lies in a neighboring cell.
-            let face_rgb = flat_color(styles.get(world.cliff_material(owner)));
-            for a in 1..6u8 {
-                if !mats.contains(&a) {
-                    continue;
-                }
-                let case = u8::from(mats[0] == a)
-                    | u8::from(mats[1] == a) << 1
-                    | u8::from(mats[2] == a) << 2
-                    | u8::from(mats[3] == a) << 3;
-                for &(p, q) in WALL_SEGMENTS[case as usize] {
-                    // The two sides of each crossed edge: the `a` (high
-                    // candidate) corner and the non-`a` (low) corner, each
-                    // read at point level, each carrying its corner sample
-                    // as the side anchor. The gate is side-aware and
-                    // per-segment — never the window center, whose side
-                    // alternates along a contour and would picket-fence the
-                    // wall with skipped windows.
-                    let sides_of = |m: u8| {
-                        let (i, j) = edge_corners(m);
-                        let (hi, lo) = if mats[i] == a { (i, j) } else { (j, i) };
-                        (
-                            mats[lo],
-                            point_level[lo],
-                            point_level[hi],
-                            corner_oct[lo],
-                            corner_oct[hi],
-                        )
-                    };
-                    let (mat_p, lvl_p, hi_p, lo_anchor_p, hi_anchor_p) = sides_of(p);
-                    let (mat_q, lvl_q, hi_q, lo_anchor_q, hi_anchor_q) = sides_of(q);
-                    let is_void = mat_p == 0 || mat_q == 0;
-                    let low_level = lvl_p.min(lvl_q);
-                    let high_level = hi_p.max(hi_q);
-                    let mp = mid(p);
-                    let mq = mid(q);
-                    let wx_p = mp[0] as f32 / OCTIMETERS_PER_METER;
-                    let wz_p = mp[1] as f32 / OCTIMETERS_PER_METER;
-                    let wx_q = mq[0] as f32 / OCTIMETERS_PER_METER;
-                    let wz_q = mq[1] as f32 / OCTIMETERS_PER_METER;
-                    // Tops anchored to each endpoint's `a`-side corner — the
-                    // identical anchor the high flap's crossing vertex took
-                    // ([`window_point_anchor`]), so the seam is shared
-                    // vertices under any smoothing displacement.
-                    let yt_p = anchored_lift(world, owner, hi_anchor_p, wx_p, wz_p);
-                    let yt_q = anchored_lift(world, owner, hi_anchor_q, wx_q, wz_q);
-                    // The low side's committed base, per endpoint. A material
-                    // boundary reads the low cap's own committed lift (over
-                    // relief the low cap splits to its own plates, so a flat
-                    // min base would gap where the low ground varies along the
-                    // segment; a relief-free owner keeps the flat min base
-                    // byte-identically). A Void low side closes to its
-                    // fill-over floor when the joint is enclosed — wall down,
-                    // floor across ([`emit_void_floors`]), wall up — coincident
-                    // with the groove floor cap; an unbounded void keeps the
-                    // open skirt.
-                    let (yb_p, yb_q) = if is_void {
-                        (
-                            void_low_base(world, lo_anchor_p, yt_p),
-                            void_low_base(world, lo_anchor_q, yt_q),
-                        )
-                    } else if relief {
-                        (
-                            anchored_lift(world, owner, lo_anchor_p, wx_p, wz_p),
-                            anchored_lift(world, owner, lo_anchor_q, wx_q, wz_q),
-                        )
-                    } else {
-                        let base = low_level as f32 / OCTIMETERS_PER_METER;
-                        (base, base)
-                    };
-                    // The split test: a face closes only where the high side
-                    // committed above the low side past the step ceiling — a
-                    // merged plate (flat ground or a legal step) closes
-                    // nothing. The solid side tests the point-level split (the
-                    // plate break); the Void side tests the committed drop to
-                    // its stored floor. The owner is only the dedup key and
-                    // color source, never the split — a low-owned window still
-                    // closes its segment.
-                    if is_void {
-                        // A Void low side lofts only where the high side stands
-                        // past the step ceiling above the void's floor: its
-                        // enclosed groove floor where authored below the
-                        // surrounding ground ([`void_floor_level`]), else the
-                        // lowest neighbor ground the skirt drops toward. A flat
-                        // void edge — the high side level with its surroundings
-                        // — lofts nothing, so a cliff-free world stays
-                        // wall-free even where grass meets an open border.
-                        let floor = void_floor_level(world, lo_anchor_p, owner)
-                            .min(void_floor_level(world, lo_anchor_q, owner));
-                        if high_level - floor <= STEP_MAX_OCTIMETERS {
-                            continue;
-                        }
-                    } else if high_level - low_level <= STEP_MAX_OCTIMETERS {
-                        continue;
-                    }
-                    push_wall_quad(
-                        tris,
-                        [wx_p, wz_p, yt_p],
-                        [wx_q, wz_q, yt_q],
-                        yb_p,
-                        yb_q,
-                        face_rgb,
-                    );
-                }
-            }
-        }
-    }
-}
-
-/// The floor a Void point grooves to — the bordering solid cell whose
-/// `cliff_material` colors it — or `None` when the point is an open
-/// silhouette edge that skirts instead. A groove is an **authored
-/// depression**: the point's [`World::point_height`] stands below its cell's
-/// base ground ([`World::height`], so a negative authored delta), *and* the
-/// joint is enclosed by solid within the fill-over reach
-/// ([`void_fill_border`]). An inherited Void point — a raised plateau's
-/// cut-away silhouette rather than a dug well — carries the cell's base
-/// height, not a floor, so it drops the skirt toward its surroundings. The
-/// floor level itself is the void point's `point_height`.
-fn void_groove_floor(world: &World, cell: CellPos, sub_x: i32, sub_z: i32) -> Option<CellPos> {
-    if world.underlay_point(cell, sub_x, sub_z) != Material::Void {
-        return None;
-    }
-    if world.point_height(cell, sub_x, sub_z) >= world.height(cell) {
-        return None; // no authored depression — an open edge, not a groove
-    }
-    let gx = cell.x * SUB + sub_x;
-    let gz = cell.z * SUB + sub_z;
-    void_fill_border(world, gx, gz)
-}
-
-/// The base a Void low side closes to at a marched segment endpoint: the
-/// void point's stored floor height ([`World::point_height`]) when the joint
-/// is an enclosed authored groove ([`void_groove_floor`]), so wall / floor /
-/// wall reads as a real flat-bottomed groove; else `yt` dropped by the
-/// unbounded-void skirt (the cut-away silhouette of a plateau, or a void that
-/// reaches the world border within the fill-over bound).
-fn void_low_base(world: &World, anchor_oct: [i32; 2], yt: f32) -> f32 {
-    let cell = CellPos {
-        x: anchor_oct[0].div_euclid(256),
-        z: anchor_oct[1].div_euclid(256),
-    };
-    let sub_x = anchor_oct[0].rem_euclid(256) / OCTIMETERS_PER_SUBCELL;
-    let sub_z = anchor_oct[1].rem_euclid(256) / OCTIMETERS_PER_SUBCELL;
-    if void_groove_floor(world, cell, sub_x, sub_z).is_some() {
-        return world.point_height(cell, sub_x, sub_z) as f32 / OCTIMETERS_PER_METER;
-    }
-    yt - WALL_VOID_SKIRT_OCTIMETERS as f32 / OCTIMETERS_PER_METER
-}
-
-/// The floor level in octimeters the Void split test reads at a segment
-/// endpoint: the groove floor of an enclosed authored depression
-/// ([`void_groove_floor`]), else the lowest of the owner's four neighbor
-/// ground levels — the surroundings the skirt drops toward. Gating the skirt
-/// on the surrounding ground (not the fixed skirt depth) is what keeps a flat
-/// void edge, where the high side sits level with its neighbors, wall-free.
-fn void_floor_level(world: &World, anchor_oct: [i32; 2], owner: CellPos) -> i32 {
-    let cell = CellPos {
-        x: anchor_oct[0].div_euclid(256),
-        z: anchor_oct[1].div_euclid(256),
-    };
-    let sub_x = anchor_oct[0].rem_euclid(256) / OCTIMETERS_PER_SUBCELL;
-    let sub_z = anchor_oct[1].rem_euclid(256) / OCTIMETERS_PER_SUBCELL;
-    if void_groove_floor(world, cell, sub_x, sub_z).is_some() {
-        return world.point_height(cell, sub_x, sub_z);
-    }
-    [(1, 0), (-1, 0), (0, 1), (0, -1)]
-        .into_iter()
-        .map(|(dx, dz)| {
-            world.surface_level(CellPos {
-                x: owner.x + dx,
-                z: owner.z + dz,
-            })
-        })
-        .min()
-        .unwrap_or_else(|| world.surface_level(owner))
-}
-
-/// The color source for a Void point's fill-over floor, or `None` when the
-/// point is not enclosed. The void point at global subcell `(gx0, gz0)` is
-/// enclosed when every one of the four axis directions reaches a solid point
-/// within [`MAX_APRON_SUBCELLS`] subcells — a bounded march past void; the
-/// returned cell is the closest such solid point's, whose `cliff_material`
-/// colors the groove. A direction that runs to the bound still in void means
-/// the void reaches the world border within reach: an open skirt, no floor.
-/// The bound is the `R = 1` remesh read reach, so the march never reads past
-/// what invalidation covers.
-fn void_fill_border(world: &World, gx0: i32, gz0: i32) -> Option<CellPos> {
-    let mut nearest: Option<(i32, CellPos)> = None;
-    for (dx, dz) in [(1, 0), (-1, 0), (0, 1), (0, -1)] {
-        let mut hit = false;
-        for step in 1..=MAX_APRON_SUBCELLS {
-            let gx = gx0 + dx * step;
-            let gz = gz0 + dz * step;
-            let cell = CellPos {
-                x: gx.div_euclid(SUB),
-                z: gz.div_euclid(SUB),
-            };
-            if world.underlay_point(cell, gx.rem_euclid(SUB), gz.rem_euclid(SUB)) != Material::Void
-            {
-                if nearest.is_none_or(|(d, _)| step < d) {
-                    nearest = Some((step, cell));
-                }
-                hit = true;
-                break;
-            }
-        }
-        if !hit {
-            return None; // open toward this side within reach — keep the skirt
-        }
-    }
-    nearest.map(|(_, cell)| cell)
-}
-
-/// Emit the fill-over floor caps: every chunk-local Void point that
-/// [`void_groove_floor`] proves an enclosed authored depression floors over at
-/// its stored point height in the bordering cell's cliff material as a flat
-/// color. The rim walls closing the groove's sides are the marched closure's
-/// Void faces, dropped to this same stored height ([`void_low_base`]), so the
-/// floor and its walls meet watertight. A cell with no floored Void point
-/// emits nothing, so a solid world — and a plateau's cut-away silhouette,
-/// whose Void points carry no authored depression — stays byte-identical.
-fn emit_void_floors(
-    world: &World,
-    at: ChunkPos,
-    styles: &StyleTable,
-    tris: &mut Vec<DrawTriangle>,
-) {
-    for lz in 0..EDGE {
-        for lx in 0..EDGE {
-            let cell = CellPos {
-                x: at.x * EDGE + lx,
-                z: at.z * EDGE + lz,
-            };
-            for sj in 0..SUB {
-                for si in 0..SUB {
-                    let Some(border) = void_groove_floor(world, cell, si, sj) else {
-                        continue; // open skirt or no authored depression — no floor
-                    };
-                    let y = world.point_height(cell, si, sj) as f32 / OCTIMETERS_PER_METER;
-                    let rgb = flat_color(styles.get(world.cliff_material(border)));
-                    let x0 = cell.x * 256 + si * OCTIMETERS_PER_SUBCELL;
-                    let z0 = cell.z * 256 + sj * OCTIMETERS_PER_SUBCELL;
-                    let vertex = |wx: f32, wz: f32| Vertex {
-                        x: wx,
-                        y,
-                        z: wz,
-                        r: rgb[0],
-                        g: rgb[1],
-                        b: rgb[2],
-                    };
-                    push_quad(
-                        tris,
-                        x0,
-                        z0,
-                        x0 + OCTIMETERS_PER_SUBCELL,
-                        z0 + OCTIMETERS_PER_SUBCELL,
-                        &vertex,
-                    );
-                }
-            }
-        }
-    }
-}
-
-/// Per wall direction: neighbor offset, the shared edge's two lattice
-/// corners as indices into the high side's corner order, and the same
-/// lattice points in the neighbor's order (the low side's plates). Used by
-/// the closure walk ([`emit_lattice_closure`]) at both the cell and subcell
-/// strides — the corner index order is the same at both.
-struct WallEdge {
-    offset: (i32, i32),
-    top: [usize; 2],
-    bottom: [usize; 2],
-}
-
-const WALL_DIRECTIONS: [WallEdge; 4] = [
-    WallEdge {
-        offset: (1, 0),
-        top: [1, 3],
-        bottom: [0, 2],
-    },
-    WallEdge {
-        offset: (-1, 0),
-        top: [0, 2],
-        bottom: [1, 3],
-    },
-    WallEdge {
-        offset: (0, 1),
-        top: [2, 3],
-        bottom: [0, 1],
-    },
-    WallEdge {
-        offset: (0, -1),
-        top: [0, 1],
-        bottom: [2, 3],
-    },
-];
-
-/// Push the two triangles of one vertical wall quad: a top edge from
-/// `top_a` to `top_b` (`[wx, wz, y]` meters) dropped to `y_bottom_a` /
-/// `y_bottom_b` at the same footprint, every vertex in the flat `color`.
-fn push_wall_quad(
-    tris: &mut Vec<DrawTriangle>,
-    top_a: [f32; 3],
-    top_b: [f32; 3],
-    y_bottom_a: f32,
-    y_bottom_b: f32,
-    color: [f32; 3],
-) {
-    let vert = |x: f32, z: f32, y: f32| Vertex {
-        x,
-        y,
-        z,
-        r: color[0],
-        g: color[1],
-        b: color[2],
-    };
-    let a = vert(top_a[0], top_a[1], top_a[2]);
-    let b = vert(top_b[0], top_b[1], top_b[2]);
-    let c = vert(top_b[0], top_b[1], y_bottom_b);
-    let d = vert(top_a[0], top_a[1], y_bottom_a);
-    tris.push(DrawTriangle { verts: [a, b, c] });
-    tris.push(DrawTriangle { verts: [a, c, d] });
-}
-
-#[cfg(test)]
-mod tests {
-    use alloc::collections::BTreeMap;
-
-    use super::*;
-    use crate::world::{
-        CELLS_PER_CHUNK_AREA, Chunk, Region, STEP_MAX_OCTIMETERS, SUBCELLS_PER_CELL, SetChunk,
-        SmoothingProfile, UNDERLAY_POINT_INHERIT,
-    };
-
-    fn sub4(n: i32) -> i32 {
-        n * SUB / 4
-    }
-
-    /// A world of one chunk whose underlay is filled per a closure.
-    fn world_with_underlay(pos: ChunkPos, fill: impl Fn(i32, i32) -> Material) -> World {
-        let mut chunk = Chunk::empty();
-        for lz in 0..EDGE {
-            for lx in 0..EDGE {
-                chunk.underlay[(lz * EDGE + lx) as usize] = fill(lx, lz);
-            }
-        }
-        let mut world = World::new();
-        world.insert_chunk(pos, chunk);
-        world
-    }
-
-    /// Two chunks of grass-default region with an explicit sand band over
-    /// cells `10..22 × 4..8` (crossing the chunk border at `x = 16`).
-    /// `profile` paints every cell's smoothing plane with the given
-    /// profile-1 settings; `None` leaves the material defaults governing.
-    fn sand_band_world(profile: Option<SmoothingProfile>) -> World {
-        let mut world = World::new();
-        world.insert_region(
-            1,
-            Region {
-                name: "meadow".into(),
-                default_material: Material::Grass,
-                cliff_material: Material::Stone,
-            },
-        );
-        if let Some(p) = profile {
-            world.insert_smoothing_profile(1, p);
-        }
-        for cx in 0..2 {
-            let mut chunk = Chunk::empty();
-            chunk.region = [1; CELLS_PER_CHUNK_AREA];
-            if profile.is_some() {
-                chunk.smoothing = [1; CELLS_PER_CHUNK_AREA];
-            }
-            for lz in 4..8 {
-                for lx in 0..EDGE {
-                    if (10..22).contains(&(cx * EDGE + lx)) {
-                        chunk.underlay[(lz * EDGE + lx) as usize] = Material::Sand;
-                    }
-                }
-            }
-            world.insert_chunk(ChunkPos { x: cx, z: 0 }, chunk);
-        }
-        world
-    }
-
-    /// Does `t` (projected to the ground plane) cover point `(px, pz)`?
-    fn covers(t: &DrawTriangle, px: f32, pz: f32) -> bool {
-        let sign = |ax: f32, az: f32, bx: f32, bz: f32| {
-            (ax - bx).mul_add(-(pz - bz), (px - bx) * (az - bz))
-        };
-        let d1 = sign(t.verts[0].x, t.verts[0].z, t.verts[1].x, t.verts[1].z);
-        let d2 = sign(t.verts[1].x, t.verts[1].z, t.verts[2].x, t.verts[2].z);
-        let d3 = sign(t.verts[2].x, t.verts[2].z, t.verts[0].x, t.verts[0].z);
-        let has_neg = d1 < -1e-6 || d2 < -1e-6 || d3 < -1e-6;
-        let has_pos = d1 > 1e-6 || d2 > 1e-6 || d3 > 1e-6;
-        !(has_neg && has_pos)
-    }
-
-    #[test]
-    fn partition_ground_has_no_gaps() {
-        // The partition must tile the painted ground exactly — every probe
-        // point strictly inside the chunk is covered by at least one
-        // ground-plane triangle. A saddle-rule or window-skip bug shows up
-        // as a hole here.
-        let world = sand_band_world(None);
-        let mesh = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        let ground: Vec<&DrawTriangle> = mesh
-            .iter()
-            .filter(|t| t.verts.iter().all(|v| v.y == 0.0))
-            .collect();
-        for j in 0..32 {
-            for i in 0..32 {
-                let px = (i as f32 + 0.37) * 0.5;
-                let pz = (j as f32 + 0.53) * 0.5;
-                assert!(
-                    ground.iter().any(|t| covers(t, px, pz)),
-                    "ground hole at ({px}, {pz})",
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn partition_chunks_tile_the_seam_together() {
-        // Every boundary window is owned by exactly one chunk (the one
-        // holding its center cell), so neither mesh alone covers the whole
-        // seam strip — but their union must, with no gap. An ownership or
-        // classification disagreement between the two chunks shows up as a
-        // hole here.
-        let world = sand_band_world(None);
-        let mesh0 = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        let mesh1 = mesh_chunk(&world, ChunkPos { x: 1, z: 0 }, &StyleTable::default());
-        let ground_covers = |mesh: &[DrawTriangle], px: f32, pz: f32| {
-            mesh.iter()
-                .filter(|t| t.verts.iter().all(|v| v.y == 0.0))
-                .any(|t| covers(t, px, pz))
-        };
-        for j in 0..64 {
-            let pz = (j as f32).mul_add(0.25, 0.13);
-            for i in 0..16 {
-                let px = (i as f32).mul_add(0.125, 15.06);
-                assert!(
-                    ground_covers(&mesh0, px, pz) || ground_covers(&mesh1, px, pz),
-                    "seam hole at ({px}, {pz})",
-                );
-            }
-        }
-    }
-
-    /// The sand-band world over a gentle eastward ramp — heights step 8
-    /// octimeters per cell, capped at the step ceiling so nothing cliffs
-    /// anywhere (a 64-octimeter drop against the unloaded zero-height
-    /// surroundings is a legal step, not a cliff).
-    fn ramp_world() -> World {
-        let mut world = sand_band_world(None);
-        for cx in 0..2 {
-            let at = ChunkPos { x: cx, z: 0 };
-            let mut chunk = world.chunk(at).expect("chunk").clone();
-            for lz in 0..EDGE {
-                for lx in 0..EDGE {
-                    let global_x = cx * EDGE + lx;
-                    chunk.height[(lz * EDGE + lx) as usize] =
-                        (8 * global_x).min(STEP_MAX_OCTIMETERS);
-                }
-            }
-            world.insert_chunk(at, chunk);
-        }
-        world
-    }
-
-    #[test]
-    fn drawn_ground_equals_stood_on_height() {
-        // The drawn-equals-stood-on tripwire: on a cliff-free ramp every
-        // vertex sits exactly on `World::surface_height` — the same patch
-        // a mover will stand on. Any lift path that misses a vertex
-        // (nine-slice, strip, window poly) diverges here.
-        let world = ramp_world();
-        let mesh = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        assert!(!mesh.is_empty());
-        for v in mesh.iter().flat_map(|t| t.verts.iter()) {
-            let surface = world.surface_height(v.x, v.z);
-            assert!(
-                (v.y - surface).abs() < 1e-4,
-                "vertex ({}, {}) drawn at {} but stood on {surface}",
-                v.x,
-                v.z,
-                v.y,
-            );
-        }
-    }
-
-    #[test]
-    fn ramp_ground_still_has_no_gaps() {
-        // The no-gap probe on sloped ground: the strip fallback and the
-        // per-window bilinear quads must tile exactly like the flat path.
-        let world = ramp_world();
-        let mesh = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        for j in 0..32 {
-            for i in 0..32 {
-                let px = (i as f32 + 0.37) * 0.5;
-                let pz = (j as f32 + 0.53) * 0.5;
-                assert!(
-                    mesh.iter().any(|t| covers(t, px, pz)),
-                    "ground hole at ({px}, {pz})",
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn a_same_material_cliff_stands_a_wall_on_the_cell_line() {
-        // A grass plateau chunk one meter up beside a grass ground chunk:
-        // same material, so the material partition leaves no marched
-        // boundary and the wall stands on the shared cell-edge line — where
-        // the old skirt stood. The face belongs to the high cell, so the
-        // high chunk lofts it exactly once and the low chunk never does;
-        // double emission would z-fight, omission would leave a window
-        // through the terrain.
-        let mut world = World::new();
-        let mut low = Chunk::empty();
-        low.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-        world.insert_chunk(ChunkPos { x: 0, z: 0 }, low);
-        let mut high = Chunk::empty();
-        high.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-        high.height = [256; CELLS_PER_CHUNK_AREA];
-        world.insert_chunk(ChunkPos { x: 1, z: 0 }, high);
-
-        let is_border_wall = |t: &DrawTriangle| {
-            t.verts.iter().all(|v| (v.x - 16.0).abs() < 1e-6)
-                && t.verts.iter().any(|v| v.y > 0.9)
-                && t.verts.iter().any(|v| v.y < 0.1)
-        };
-        let low_mesh = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        let high_mesh = mesh_chunk(&world, ChunkPos { x: 1, z: 0 }, &StyleTable::default());
-        assert!(
-            high_mesh.iter().any(is_border_wall),
-            "the high chunk stands the wall on the cell line",
-        );
-        assert!(
-            !low_mesh.iter().any(is_border_wall),
-            "the low chunk does not double-draw the shared face",
-        );
-    }
-
-    #[test]
-    fn a_material_boundary_at_a_cliff_does_not_drape() {
-        // Sand plateau (1 m up) against grass ground, material boundary on
-        // the cliff line: every colored (non-wall) triangle must stay on
-        // its own side of the break — the cliff face belongs to the gray
-        // wall. Position-pure lifting would stretch the boundary windows
-        // into meter-tall sand and grass caps down the face.
-        let mut world = World::new();
-        let mut low = Chunk::empty();
-        low.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-        let mut high = Chunk::empty();
-        high.underlay = [Material::Sand; CELLS_PER_CHUNK_AREA];
-        high.height = [256; CELLS_PER_CHUNK_AREA];
-        world.insert_chunk(ChunkPos { x: 0, z: 0 }, low);
-        world.insert_chunk(ChunkPos { x: 1, z: 0 }, high);
-
-        for at in [ChunkPos { x: 0, z: 0 }, ChunkPos { x: 1, z: 0 }] {
-            let mesh = mesh_chunk(&world, at, &StyleTable::default());
-            for t in &mesh {
-                let gray = t
-                    .verts
-                    .iter()
-                    .all(|v| (v.r - v.b).abs() < 0.05 && (v.g - v.b).abs() < 0.05);
-                if gray {
-                    continue; // walls own the vertical face
-                }
-                let max_y = t.verts.iter().map(|v| v.y).fold(f32::MIN, f32::max);
-                let min_y = t.verts.iter().map(|v| v.y).fold(f32::MAX, f32::min);
-                assert!(
-                    max_y - min_y < 0.5,
-                    "chunk {at:?}: a colored triangle drapes the cliff (span {} at x {})",
-                    max_y - min_y,
-                    t.verts[0].x,
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn void_chunk_meshes_to_nothing() {
-        let mut world = World::new();
-        world.insert_chunk(ChunkPos { x: 0, z: 0 }, Chunk::empty());
-        let tris = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        assert!(tris.is_empty(), "an all-Void chunk emits no geometry");
-    }
-
-    #[test]
-    fn overlay_helpers_collect_materials_and_gate_coverage() {
-        let at = ChunkPos { x: 0, z: 0 };
-        let mut chunk = Chunk::empty();
-        let stone_cell = (8 * EDGE + 8) as usize;
-        let grass_cell = (2 * EDGE + 2) as usize;
-        chunk.overlay[stone_cell] = Material::Stone;
-        chunk.overlay[grass_cell] = Material::Grass;
-        chunk.overlay_mask[stone_cell * SUBCELLS_PER_CELL + 3] = 77;
-        let mut world = World::new();
-        world.insert_chunk(at, chunk);
-
-        assert_eq!(
-            overlay_materials(&world, at),
-            vec![Material::Grass, Material::Stone],
-            "overlay materials are distinct and stable by material id",
-        );
-        assert_eq!(
-            subcell_coverage(&world, at, 8 * SUB + 3, 8 * SUB, Material::Stone),
-            77,
-            "matching overlay material exposes its scalar mask sample",
-        );
-        assert_eq!(
-            subcell_coverage(&world, at, 8 * SUB + 3, 8 * SUB, Material::Grass),
-            0,
-            "off-material coverage is treated as uncovered",
-        );
-    }
-
-    #[test]
-    fn scalar_overlay_mask_marches_at_authored_fraction() {
-        // One authored overlay cell carries a scalar vertical edge. The
-        // contour pass first reconstructs scalar zones bilinearly, then
-        // marches the 127.5 crossing through that reconstructed grid. A
-        // nearest/blocky path would land at the half-cell edge instead.
-        let at = ChunkPos { x: 0, z: 0 };
-        let cell = CellPos { x: 8, z: 8 };
-        let cell_idx = (cell.z * EDGE + cell.x) as usize;
-        let sub = SUB as usize;
-        let mut overlay = vec![Material::Void.to_u8(); CELLS_PER_CHUNK_AREA];
-        overlay[cell_idx] = Material::Stone.to_u8();
-        let mut overlay_mask = vec![0u8; CELLS_PER_CHUNK_AREA * SUBCELLS_PER_CELL];
-        for sz in 0..sub {
-            for sx in 0..sub {
-                overlay_mask[cell_idx * SUBCELLS_PER_CELL + sz * sub + sx] =
-                    if sx < sub / 2 { 96 } else { 192 };
-            }
-        }
-        let mut world = World::new();
-        world.insert_chunk(
-            at,
-            SetChunk {
-                chunk_x: 0,
-                chunk_z: 0,
-                underlay: vec![Material::Grass.to_u8(); CELLS_PER_CHUNK_AREA],
-                underlay_points: Vec::new(),
-                height_points: Vec::new(),
-                overlay,
-                overlay_mask,
-                height: Vec::new(),
-                region: Vec::new(),
-                water_plane: Vec::new(),
-                smoothing: Vec::new(),
-            }
-            .into_chunk(),
-        );
-
-        let mesh = mesh_chunk(&world, at, &StyleTable::default());
-        let coverage_verts: Vec<&Vertex> = mesh
-            .iter()
-            .filter(|t| t.verts.iter().all(|v| (v.y - COVERAGE_LIFT).abs() < 1e-6))
-            .flat_map(|t| t.verts.iter())
-            .collect();
-        assert!(
-            !coverage_verts.is_empty(),
-            "the overlay coverage pass emits lifted contour geometry",
-        );
-
-        let low_sample_x = cell.x as f32 + (sub / 2 - 2) as f32 / SUB as f32 + 0.5 / SUB as f32;
-        let high_sample_x = cell.x as f32 + (sub / 2 - 1) as f32 / SUB as f32 + 0.5 / SUB as f32;
-        let low_reconstructed = 96.0;
-        let high_reconstructed = (96.0 + 192.0) * 0.5;
-        let crossing_t = (127.5 - low_reconstructed) / (high_reconstructed - low_reconstructed);
-        let expected_x = low_sample_x + (high_sample_x - low_sample_x) * crossing_t;
-        let min_x = coverage_verts.iter().map(|v| v.x).fold(f32::MAX, f32::min);
-        assert!(
-            (min_x - expected_x).abs() < 1e-4,
-            "scalar contour crosses at {expected_x}, got {min_x}",
-        );
-        assert!(
-            (min_x - (cell.x as f32 + 0.5)).abs() > 1e-3,
-            "the contour did not collapse to the blocky half-cell edge",
-        );
-    }
-
-    #[test]
-    fn all_inherit_meshes_identically_to_cell_expansion() {
-        // Tripwire: the underlay-point inherit sentinel must fold back to the
-        // per-cell material exactly, so an unshaped world meshes byte-for-byte
-        // as it did before points existed. A varied world meshed with
-        // all-inherit points must equal the same world with every point
-        // explicitly pinned to its cell's cascade material — the literal
-        // cell-expansion `partition_inputs` performed before. Any drift in
-        // `underlay_point`'s indexing or sentinel handling breaks this.
-        let at = ChunkPos { x: 0, z: 0 };
-        let mut world = World::new();
-        let mut chunk = Chunk::empty();
-        for lz in 0..EDGE {
-            for lx in 0..EDGE {
-                let i = (lz * EDGE + lx) as usize;
-                chunk.underlay[i] = match (lx + lz) % 4 {
-                    0 => Material::Void, // falls through to the region default
-                    1 => Material::Grass,
-                    2 => Material::Sand,
-                    _ => Material::Stone,
-                };
-                chunk.height[i] = 10 * lx;
-                chunk.region[i] = 1;
-            }
-        }
-        world.insert_chunk(at, chunk);
-        world.insert_region(
-            1,
-            Region {
-                name: "meadow".into(),
-                default_material: Material::Dirt,
-                cliff_material: Material::Stone,
-            },
-        );
-
-        let inherit_mesh = mesh_chunk(&world, at, &StyleTable::default());
-        assert!(!inherit_mesh.is_empty(), "the fixture must mesh something");
-
-        // Pin every cell's points to its own cascade material — the explicit
-        // cell-expansion. Apron cells stay off-chunk and read Void either way.
-        let mut expanded = world.clone();
-        for lz in 0..EDGE {
-            for lx in 0..EDGE {
-                let cell = CellPos {
-                    x: at.x * EDGE + lx,
-                    z: at.z * EDGE + lz,
-                };
-                let material = expanded.underlay(cell).to_u8();
-                expanded.set_cell_points(cell, &[material; SUBCELLS_PER_CELL]);
-            }
-        }
-        let expanded_mesh = mesh_chunk(&expanded, at, &StyleTable::default());
-
-        assert_eq!(
-            inherit_mesh, expanded_mesh,
-            "all-inherit points fold to the explicit cell-expansion",
-        );
-    }
-
-    #[test]
-    fn a_point_shaped_cell_marches_inside_its_cell_span() {
-        // A cell whose only authored material is the middle half-cell of
-        // interior subcells (the rest inherit Void) marches a silhouette that
-        // departs the cell-edge lines: the Grass/Void contour lands between
-        // subcell centers, strictly inside the cell span, where a whole-cell
-        // material would only bound at the integer cell edges.
-        let at = ChunkPos { x: 0, z: 0 };
-        let cell = CellPos { x: 8, z: 8 };
-        let sub = SUB as usize;
-        let mut points = [UNDERLAY_POINT_INHERIT; SUBCELLS_PER_CELL];
-        for sub_z in sub / 4..3 * sub / 4 {
-            for sub_x in sub / 4..3 * sub / 4 {
-                points[sub_z * sub + sub_x] = Material::Grass.to_u8();
-            }
-        }
-        let mut world = World::new();
-        world.insert_chunk(at, Chunk::empty()); // all Void underlay
-        world.set_cell_points(cell, &points);
-
-        let mesh = mesh_chunk(&world, at, &StyleTable::default());
-        assert!(!mesh.is_empty(), "the authored blob must mesh");
-
-        let verts = mesh.iter().flat_map(|t| t.verts.iter());
-        let inside_x = verts
-            .clone()
-            .any(|v| v.x > cell.x as f32 + 0.1 && v.x < cell.x as f32 + 0.9);
-        let inside_z = verts
-            .clone()
-            .any(|v| v.z > cell.z as f32 + 0.1 && v.z < cell.z as f32 + 0.9);
-        assert!(
-            inside_x && inside_z,
-            "the marched contour must depart the cell-edge lines",
-        );
-        // And it stays bounded inside the cell — no vertex escapes the span.
-        assert!(
-            verts.clone().all(|v| {
-                v.x >= cell.x as f32 - 1e-4
-                    && v.x <= cell.x as f32 + 1.0 + 1e-4
-                    && v.z >= cell.z as f32 - 1e-4
-                    && v.z <= cell.z as f32 + 1.0 + 1e-4
-            }),
-            "the shaped silhouette stays inside the cell",
-        );
-    }
-
-    /// The vertical span of a triangle — a wall face stands ~1 m tall while
-    /// a cap or ground quad is near-flat, so the split filters walls from
-    /// the surfaces they join.
-    fn y_span(t: &DrawTriangle) -> f32 {
-        let max = t.verts.iter().map(|v| v.y).fold(f32::MIN, f32::max);
-        let min = t.verts.iter().map(|v| v.y).fold(f32::MAX, f32::min);
-        max - min
-    }
-
-    #[test]
-    fn marched_wall_tops_land_on_the_cap_contour() {
-        // The watertight pin: a material-boundary cliff lofts marched walls
-        // whose top vertices are the cap contour's own vertices. Sand plateau
-        // (1 m up, east) against grass ground: mesh the high chunk, split its
-        // triangles into walls (a tall vertical span) and caps (near-flat),
-        // and assert every wall-top vertex coincides exactly with a cap
-        // vertex — no gap and no overlap between the cliff face and the
-        // surfaces it joins. The wall lofts its top through the same
-        // owner-clamped patch the cap drew, so the match is bit-exact.
-        let mut world = World::new();
-        let mut low = Chunk::empty();
-        low.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-        let mut high = Chunk::empty();
-        high.underlay = [Material::Sand; CELLS_PER_CHUNK_AREA];
-        high.height = [256; CELLS_PER_CHUNK_AREA];
-        world.insert_chunk(ChunkPos { x: 0, z: 0 }, low);
-        world.insert_chunk(ChunkPos { x: 1, z: 0 }, high);
-
-        let mesh = mesh_chunk(&world, ChunkPos { x: 1, z: 0 }, &StyleTable::default());
-        let walls: Vec<&DrawTriangle> = mesh.iter().filter(|t| y_span(t) > 0.5).collect();
-        assert!(
-            !walls.is_empty(),
-            "the material-boundary cliff lofts marched walls",
-        );
-        let cap_verts: Vec<&Vertex> = mesh
-            .iter()
-            .filter(|t| y_span(t) <= 0.5)
-            .flat_map(|t| t.verts.iter())
-            .collect();
-        for wall in &walls {
-            let top = wall.verts.iter().map(|v| v.y).fold(f32::MIN, f32::max);
-            for v in wall.verts.iter().filter(|v| (v.y - top).abs() < 1e-6) {
-                assert!(
-                    cap_verts.iter().any(|c| (c.x - v.x).abs() < 1e-6
-                        && (c.y - v.y).abs() < 1e-6
-                        && (c.z - v.z).abs() < 1e-6),
-                    "wall-top vertex ({}, {}, {}) has no matching cap vertex",
-                    v.x,
-                    v.y,
-                    v.z,
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn a_flat_world_emits_no_walls() {
-        // Tripwire: a cliff-free world lofts no walls, so the wall pass adds
-        // nothing — the full mesh equals the underlay mesh built without it,
-        // byte for byte. A wall pass that fired on flat ground (or
-        // double-counted a lattice against a marched segment) diverges here.
-        let world = world_with_underlay(ChunkPos { x: 0, z: 0 }, |_, _| Material::Grass);
-        let at = ChunkPos { x: 0, z: 0 };
-        let styles = StyleTable::default();
-        let mut expected = Vec::new();
-        let _ = mesh_underlay(&world, at, &styles, &mut expected);
-        let full = mesh_chunk(&world, at, &styles);
-        assert_eq!(full, expected, "the wall pass adds nothing to a flat world");
-    }
-
-    #[test]
-    fn adjacent_chunks_agree_on_a_wall_across_their_seam() {
-        // A grass cliff running in x across the vertical chunk seam at
-        // x = 16: cells at z < 8 sit 1 m up, z >= 8 at the datum, same
-        // material so the face is a cell-edge lattice wall. Each chunk owns
-        // its half of the line; at the shared seam the two halves must meet
-        // on identical vertices, or the wall would gap or double where the
-        // chunks join.
-        let mut world = World::new();
-        for cx in 0..2 {
-            let mut chunk = Chunk::empty();
-            chunk.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-            for lz in 0..EDGE {
-                for lx in 0..EDGE {
-                    chunk.height[(lz * EDGE + lx) as usize] = if lz < 8 { 256 } else { 0 };
-                }
-            }
-            world.insert_chunk(ChunkPos { x: cx, z: 0 }, chunk);
-        }
-        let seam_verts = |at| {
-            let mesh = mesh_chunk(&world, at, &StyleTable::default());
-            let mut verts: Vec<(i64, i64, i64)> = mesh
-                .iter()
-                .filter(|t| y_span(t) > 0.5)
-                .flat_map(|t| t.verts.iter())
-                .filter(|v| (v.x - 16.0).abs() < 1e-4)
-                .map(|v| {
-                    (
-                        (v.x * 256.0).round() as i64,
-                        (v.y * 256.0).round() as i64,
-                        (v.z * 256.0).round() as i64,
-                    )
-                })
-                .collect();
-            verts.sort_unstable();
-            verts.dedup();
-            verts
-        };
-        let west = seam_verts(ChunkPos { x: 0, z: 0 });
-        let east = seam_verts(ChunkPos { x: 1, z: 0 });
-        assert!(!west.is_empty(), "the seam carries a wall");
-        assert_eq!(
-            west, east,
-            "both chunks place the seam wall on identical vertices",
-        );
-    }
-
-    #[test]
-    fn a_raised_point_cell_wears_marched_walls() {
-        // A point-authored blob raised to a plateau wears walls that follow
-        // its marched silhouette, not the cell edge: a half-cell grass point
-        // block in one grass cell (the rest Void points) lifted 2 m stands a
-        // Void-drop wall whose top vertices sit strictly inside the cell
-        // span, where its Grass/Void contour marches — the cap-detaches-
-        // from-the-cell-edge case the cell-edge skirt could not close.
-        let at = ChunkPos { x: 0, z: 0 };
-        let cell = CellPos { x: 8, z: 8 };
-        let sub = SUB as usize;
-        // Explicit Void points around a half-cell grass core, so the raised
-        // cell's grass silhouette pulls inside its own edges.
-        let mut points = [Material::Void.to_u8(); SUBCELLS_PER_CELL];
-        for sub_z in sub / 4..3 * sub / 4 {
-            for sub_x in sub / 4..3 * sub / 4 {
-                points[sub_z * sub + sub_x] = Material::Grass.to_u8();
-            }
-        }
-        // Surround the raised cell with lower grass ground so the blob is a
-        // genuine plateau; the cell itself is grass at 512 octimeters.
-        let mut chunk = Chunk::empty();
-        chunk.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-        chunk.height[(cell.z * EDGE + cell.x) as usize] = 512;
-        let mut world = World::new();
-        world.insert_chunk(at, chunk);
-        world.set_cell_points(cell, &points);
-
-        let mesh = mesh_chunk(&world, at, &StyleTable::default());
-        let walls: Vec<&DrawTriangle> = mesh.iter().filter(|t| y_span(t) > 0.5).collect();
-        assert!(!walls.is_empty(), "the raised blob wears walls");
-        let inside = walls.iter().flat_map(|t| t.verts.iter()).any(|v| {
-            v.x > cell.x as f32 + 0.1
-                && v.x < cell.x as f32 + 0.9
-                && v.z > cell.z as f32 + 0.1
-                && v.z < cell.z as f32 + 0.9
-        });
-        assert!(
-            inside,
-            "the walls follow the marched silhouette inside the cell",
-        );
-    }
-
-    #[test]
-    fn drawn_relief_equals_stood_on() {
-        // drawn≡stood-on over authored subcell relief: a continuous per-point
-        // hill (deltas within the step ceiling, returning to zero at the chunk
-        // edges) lofts no walls, and every cap vertex sits exactly on
-        // `World::surface_height` — the same subcell patch a mover stands on.
-        let at = ChunkPos { x: 0, z: 0 };
-        let mut world = World::new();
-        let mut chunk = Chunk::empty();
-        chunk.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-        world.insert_chunk(at, chunk);
-        let sub = SUB as usize;
-        // A pyramid: 12 octimeters per subcell up to the middle and back down
-        // on both axes, flat (zero) by every chunk edge so the grass never
-        // stands above the Void surround and nothing cliffs anywhere.
-        let density_scale = SUB / 4;
-        let ramp = |g: i32| (12 * g.min(sub4(60) - g).max(0) / density_scale) as i16;
-        for lz in 0..EDGE {
-            for lx in 0..EDGE {
-                let mut deltas = [0i16; SUBCELLS_PER_CELL];
-                for sz in 0..sub {
-                    for sx in 0..sub {
-                        let dx = ramp(lx * SUB + sx as i32);
-                        let dz = ramp(lz * SUB + sz as i32);
-                        deltas[sz * sub + sx] = dx.min(dz);
-                    }
-                }
-                world.set_cell_heights(CellPos { x: lx, z: lz }, &deltas);
-            }
-        }
-        let mesh = mesh_chunk(&world, at, &StyleTable::default());
-        assert!(!mesh.is_empty(), "the relief cell meshes caps");
-        assert!(
-            mesh.iter().all(|t| y_span(t) < 0.5),
-            "a continuous hill lofts no walls",
-        );
-        let mut raised = false;
-        for v in mesh.iter().flat_map(|t| t.verts.iter()) {
-            let surface = world.surface_height(v.x, v.z);
-            assert!(
-                (v.y - surface).abs() < 2e-3,
-                "cap vertex ({}, {}) drawn at {} but stood on {surface}",
-                v.x,
-                v.z,
-                v.y,
-            );
-            raised |= v.y > 1.0; // the ridge peaks near 360/256 ≈ 1.4 m
-        }
-        assert!(raised, "the authored relief actually lifts the cap");
-    }
-
-    /// Author a stone diamond (rotated square) at point resolution centered on
-    /// subcell `(34, 34)` with subcell radius `r`, raised `lift` octimeters on
-    /// a Void surround; the diamond's edges cross cell interiors diagonally.
-    fn diamond_world(r: i32, lift: i16) -> World {
-        let at = ChunkPos { x: 0, z: 0 };
-        let mut world = World::new();
-        world.insert_chunk(at, Chunk::empty()); // all-Void underlay
-        let sub = SUB as usize;
-        for lz in 0..EDGE {
-            for lx in 0..EDGE {
-                let mut points = [Material::Void.to_u8(); SUBCELLS_PER_CELL];
-                let mut deltas = [0i16; SUBCELLS_PER_CELL];
-                for sz in 0..sub {
-                    for sx in 0..sub {
-                        let gx = lx * SUB + sx as i32;
-                        let gz = lz * SUB + sz as i32;
-                        if (gx - sub4(34)).abs() + (gz - sub4(34)).abs() <= sub4(r) {
-                            points[sz * sub + sx] = Material::Stone.to_u8();
-                            deltas[sz * sub + sx] = lift;
-                        }
-                    }
-                }
-                let cell = CellPos { x: lx, z: lz };
-                world.set_cell_points(cell, &points);
-                world.set_cell_heights(cell, &deltas);
-            }
-        }
-        world
-    }
-
-    #[test]
-    fn an_authored_diamond_wears_walls_on_its_silhouette() {
-        // A point-authored stone diamond raised 300 octimeters wears walls on
-        // its marched silhouette (which crosses cell interiors diagonally),
-        // with no wall on an interior cell edge between two fully-raised cells.
-        let at = ChunkPos { x: 0, z: 0 };
-        let world = diamond_world(12, 300);
-        let mesh = mesh_chunk(&world, at, &StyleTable::default());
-        let walls: Vec<&DrawTriangle> = mesh.iter().filter(|t| y_span(t) > 0.5).collect();
-        assert!(!walls.is_empty(), "the raised diamond wears walls");
-
-        // The silhouette departs the cell grid: some wall vertex sits strictly
-        // inside a cell, off every integer lattice line.
-        let departs = walls.iter().flat_map(|t| t.verts.iter()).any(|v| {
-            let fx = v.x - v.x.floor();
-            let fz = v.z - v.z.floor();
-            (0.1..0.9).contains(&fx) && (0.1..0.9).contains(&fz)
-        });
-        assert!(
-            departs,
-            "walls follow the diagonal silhouette, not cell edges"
-        );
-
-        // No wall on the interior cell edge x = 8 between cells (7,8) and
-        // (8,8) — both fully inside the diamond (the boundary crosses x = 8
-        // only at z = 6 and z = 11), so a fused interior stands no wall.
-        let interior_edge = walls
-            .iter()
-            .flat_map(|t| t.verts.iter())
-            .any(|v| (v.x - 8.0).abs() < 1e-4 && (8.05..8.95).contains(&v.z));
-        assert!(
-            !interior_edge,
-            "no wall vertex sits on an interior cell edge inside the diamond",
-        );
-
-        // And the cap actually rides the raised plane.
-        let peak = mesh
-            .iter()
-            .filter(|t| y_span(t) < 0.5)
-            .flat_map(|t| t.verts.iter())
-            .map(|v| v.y)
-            .fold(f32::MIN, f32::max);
-        assert!(
-            (peak - 300.0 / 256.0).abs() < 1e-3,
-            "the diamond cap stands at the raised level, peak {peak}",
-        );
-    }
-
-    #[test]
-    fn fused_equal_height_points_stand_no_internal_wall() {
-        // Two stone cells raised to the same per-point level fuse: the wall
-        // pass closes their outer perimeter against the Void surround but
-        // stands no wall on their shared edge, where the heights match.
-        let at = ChunkPos { x: 0, z: 0 };
-        let mut world = World::new();
-        let mut chunk = Chunk::empty();
-        chunk.underlay[8 * EDGE as usize + 8] = Material::Stone; // cell (8,8)
-        chunk.underlay[8 * EDGE as usize + 9] = Material::Stone; // cell (9,8)
-        world.insert_chunk(at, chunk);
-        world.set_cell_heights(CellPos { x: 8, z: 8 }, &[300; SUBCELLS_PER_CELL]);
-        world.set_cell_heights(CellPos { x: 9, z: 8 }, &[300; SUBCELLS_PER_CELL]);
-
-        let mesh = mesh_chunk(&world, at, &StyleTable::default());
-        let walls: Vec<&DrawTriangle> = mesh.iter().filter(|t| y_span(t) > 0.5).collect();
-        assert!(
-            !walls.is_empty(),
-            "the fused pair wears an outer perimeter wall"
-        );
-        let shared_edge = walls
-            .iter()
-            .flat_map(|t| t.verts.iter())
-            .any(|v| (v.x - 9.0).abs() < 1e-4 && (8.05..8.95).contains(&v.z));
-        assert!(
-            !shared_edge,
-            "no wall stands on the fused pair's equal-height shared edge",
-        );
-    }
-
-    #[test]
-    fn adjacent_chunks_agree_on_an_authored_break() {
-        // A stone shelf (raised 300, rows z < 8) meets grass ground (rows
-        // z >= 8) along a break running in x across the vertical chunk seam at
-        // x = 16. The marched wall pass reads point levels — pure functions of
-        // world position — so each chunk lofts its half and the two meet on
-        // identical vertices at the shared seam column.
-        let mut world = World::new();
-        for cx in 0..2 {
-            let mut chunk = Chunk::empty();
-            for lz in 0..EDGE {
-                for lx in 0..EDGE {
-                    let i = (lz * EDGE + lx) as usize;
-                    chunk.underlay[i] = if lz < 8 {
-                        Material::Stone
-                    } else {
-                        Material::Grass
-                    };
-                }
-            }
-            world.insert_chunk(ChunkPos { x: cx, z: 0 }, chunk);
-        }
-        for cx in 0..2 {
-            for lz in 0..8 {
-                for lx in 0..EDGE {
-                    world.set_cell_heights(
-                        CellPos {
-                            x: cx * EDGE + lx,
-                            z: lz,
-                        },
-                        &[300; SUBCELLS_PER_CELL],
-                    );
-                }
-            }
-        }
-        // Marched wall vertices land on subcell sample lines, not the exact
-        // chunk edge, so agreement is that the two chunks emit an identical
-        // vertex where their seam-straddling windows share an edge — a
-        // non-empty intersection of their wall-vertex sets, near the seam.
-        // The z band pins the collection to the authored z = 8 break line
-        // (the shelf's outer Void-drop walls are a different feature).
-        let wall_verts = |at| {
-            let mesh = mesh_chunk(&world, at, &StyleTable::default());
-            let mut verts: Vec<(i64, i64, i64)> = mesh
-                .iter()
-                .filter(|t| y_span(t) > 0.5)
-                .flat_map(|t| t.verts.iter())
-                .filter(|v| (7.5..8.5).contains(&v.z))
-                .map(|v| {
-                    (
-                        (v.x * 256.0).round() as i64,
-                        (v.y * 256.0).round() as i64,
-                        (v.z * 256.0).round() as i64,
-                    )
-                })
-                .collect();
-            verts.sort_unstable();
-            verts.dedup();
-            verts
-        };
-        let west = wall_verts(ChunkPos { x: 0, z: 0 });
-        let east = wall_verts(ChunkPos { x: 1, z: 0 });
-        assert!(
-            !west.is_empty() && !east.is_empty(),
-            "each chunk lofts the break"
-        );
-        let shared: Vec<_> = west.iter().filter(|v| east.contains(v)).collect();
-        assert!(
-            !shared.is_empty(),
-            "the two chunks meet on identical seam vertices",
-        );
-        // The shared vertices sit at the seam column and the raised top level.
-        assert!(
-            shared
-                .iter()
-                .any(|v| (v.0 - 16 * 256).abs() <= 32 && v.1 == 300),
-            "the shared seam wall stands at the authored break level",
-        );
-    }
-
-    #[test]
-    fn a_pure_delta_plateau_stands_walls_on_its_break_lines() {
-        // The same-material height-break class: a stone field at uniform cell
-        // height wears a plateau authored purely with point deltas — no
-        // material boundary anywhere (nothing for the marched pass) and no
-        // cell-height cliff (nothing for the cell-edge walk) — so every wall
-        // must come from the subcell relief walk, standing exactly on the
-        // plateau's break lines. Those lines run through cell interiors
-        // (x and z = 7.5 / 9.5 m), where a cell-edge wall cannot stand.
-        let at = ChunkPos { x: 0, z: 0 };
-        let mut world = World::new();
-        for dz in -1..=1 {
-            for dx in -1..=1 {
-                let mut c = Chunk::empty();
-                c.underlay = [Material::Stone; CELLS_PER_CHUNK_AREA];
-                world.insert_chunk(ChunkPos { x: dx, z: dz }, c);
-            }
-        }
-        // Raise the v2.5 global subcell span [30, 38) x [30, 38), scaled to
-        // the current density, by 300 octimeters.
-        let lo_sub = sub4(30);
-        let hi_sub = sub4(38);
-        let sub = SUB as usize;
-        for lz in 6..10 {
-            for lx in 6..10 {
-                let mut deltas = [0i16; SUBCELLS_PER_CELL];
-                for sz in 0..sub {
-                    for sx in 0..sub {
-                        let gx = lx * SUB + sx as i32;
-                        let gz = lz * SUB + sz as i32;
-                        if (lo_sub..hi_sub).contains(&gx) && (lo_sub..hi_sub).contains(&gz) {
-                            deltas[sz * sub + sx] = 300;
-                        }
-                    }
-                }
-                world.set_cell_heights(CellPos { x: lx, z: lz }, &deltas);
-            }
-        }
-        let mesh = mesh_chunk(&world, at, &StyleTable::default());
-        let walls: Vec<&DrawTriangle> = mesh.iter().filter(|t| y_span(t) > 0.5).collect();
-        assert!(!walls.is_empty(), "the delta plateau wears walls");
-        // Every wall vertex lies on the block's perimeter lines — the
-        // authored break, nowhere else (no fused-interior or cell-edge wall).
-        let lo = 7.5f32;
-        let hi = 9.5f32;
-        for v in walls.iter().flat_map(|t| t.verts.iter()) {
-            let on_x = ((v.x - lo).abs() < 1e-4 || (v.x - hi).abs() < 1e-4)
-                && (lo - 1e-4..=hi + 1e-4).contains(&v.z);
-            let on_z = ((v.z - lo).abs() < 1e-4 || (v.z - hi).abs() < 1e-4)
-                && (lo - 1e-4..=hi + 1e-4).contains(&v.x);
-            assert!(
-                on_x || on_z,
-                "wall vertex ({}, {}) stands off the authored break lines",
-                v.x,
-                v.z,
-            );
-        }
-        // And all four perimeter sides are closed.
-        for (side, pick) in [
-            (
-                "west",
-                &(|v: &Vertex| (v.x - lo).abs() < 1e-4) as &dyn Fn(&Vertex) -> bool,
-            ),
-            ("east", &|v: &Vertex| (v.x - hi).abs() < 1e-4),
-            ("north", &|v: &Vertex| (v.z - lo).abs() < 1e-4),
-            ("south", &|v: &Vertex| (v.z - hi).abs() < 1e-4),
-        ] {
-            assert!(
-                walls.iter().flat_map(|t| t.verts.iter()).any(pick),
-                "the plateau's {side} side is open",
-            );
-        }
-    }
-
-    #[test]
-    fn a_relief_cell_on_a_cell_cliff_lofts_its_wall_exactly_once() {
-        // A relief cell standing a plain cell-height cliff above a
-        // same-material neighbor: the cell-edge walk skips relief cells and
-        // the subcell relief walk owns the face, so the shared edge must be
-        // covered exactly once — the wall segments' top edges sum to the one
-        // cell-edge meter. Double emission (both walks firing) would sum to
-        // two; a routing hole would sum to zero.
-        let at = ChunkPos { x: 0, z: 0 };
-        let mut world = World::new();
-        let mut chunk = Chunk::empty();
-        chunk.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-        chunk.height[(8 * EDGE + 8) as usize] = 256; // cell (8,8) one meter up
-        world.insert_chunk(at, chunk);
-        // A small legal-step delta inside the raised cell flips it (and its
-        // neighbors) onto the point path without adding any further break.
-        let mut deltas = [0i16; SUBCELLS_PER_CELL];
-        deltas[SUBCELLS_PER_CELL - 1] = 16;
-        world.set_cell_heights(CellPos { x: 8, z: 8 }, &deltas);
-
-        let mesh = mesh_chunk(&world, at, &StyleTable::default());
-        // Wall triangles on the west edge plane x = 8 over cell (8,8)'s span.
-        let west: Vec<&DrawTriangle> = mesh
-            .iter()
-            .filter(|t| {
-                y_span(t) > 0.5
-                    && t.verts.iter().all(|v| {
-                        (v.x - 8.0).abs() < 1e-4 && (8.0 - 1e-4..=9.0 + 1e-4).contains(&v.z)
-                    })
-            })
-            .collect();
-        assert!(!west.is_empty(), "the cliff edge carries a wall");
-        // Each wall quad's top edge appears on exactly one of its two
-        // triangles (the one holding both top vertices); summing those top
-        // spans measures the edge coverage.
-        let top_length: f32 = west
-            .iter()
-            .map(|t| {
-                let tops: Vec<&Vertex> = t.verts.iter().filter(|v| v.y > 0.5).collect();
-                if tops.len() == 2 {
-                    (tops[0].z - tops[1].z).abs()
-                } else {
-                    0.0
-                }
-            })
-            .sum();
-        assert!(
-            (top_length - 1.0).abs() < 1e-3,
-            "the shared edge is covered exactly once, got {top_length} m of top edge",
-        );
-    }
-
-    /// The authored-silhouette fixture from the causeway demo failure: a
-    /// stone square raised purely by point deltas over flat grass ground —
-    /// materials and heights break on the same subcell lines (global
-    /// subcells `[30, 38)²`, i.e. `[7.5, 9.5)` m), every cell height zero.
-    /// Smoothing is pinned off through the per-cell override plane so the
-    /// marched contour is exactly the authored break line.
-    fn delta_column_world() -> World {
-        let mut world = World::new();
-        world.insert_smoothing_profile(
-            1,
-            SmoothingProfile {
-                iterations: 0,
-                degrees: 90,
-            },
-        );
-        for dz in -1..=1 {
-            for dx in -1..=1 {
-                let mut c = Chunk::empty();
-                c.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-                c.smoothing = [1; CELLS_PER_CHUNK_AREA];
-                world.insert_chunk(ChunkPos { x: dx, z: dz }, c);
-            }
-        }
-        let sub = SUB as usize;
-        let lo_sub = sub4(30);
-        let hi_sub = sub4(38);
-        for lz in 6..10 {
-            for lx in 6..10 {
-                let mut points = [UNDERLAY_POINT_INHERIT; SUBCELLS_PER_CELL];
-                let mut deltas = [0i16; SUBCELLS_PER_CELL];
-                for sz in 0..sub {
-                    for sx in 0..sub {
-                        let gx = lx * SUB + sx as i32;
-                        let gz = lz * SUB + sz as i32;
-                        if (lo_sub..hi_sub).contains(&gx) && (lo_sub..hi_sub).contains(&gz) {
-                            points[sz * sub + sx] = Material::Stone.to_u8();
-                            deltas[sz * sub + sx] = 300;
-                        }
-                    }
-                }
-                let cell = CellPos { x: lx, z: lz };
-                world.set_cell_points(cell, &points);
-                world.set_cell_heights(cell, &deltas);
-            }
-        }
-        world
-    }
-
-    /// A cell-aligned stone block raised a whole-cell step above flat grass —
-    /// a MATERIAL boundary with **no subcell relief**, the relief-free cliff
-    /// that [`delta_column_world`]'s subcell deltas never exercise. Whole-cell
-    /// `height` only (`height_points` left inheriting), so
-    /// `cell_has_height_relief` stays false across the boundary. Smoothing is
-    /// pinned off (profile 1, iterations 0) so the marched contour stays
-    /// axis-aligned and the block's perimeter walls run straight.
-    fn material_block_world() -> World {
-        let mut world = World::new();
-        world.insert_smoothing_profile(
-            1,
-            SmoothingProfile {
-                iterations: 0,
-                degrees: 90,
-            },
-        );
-        for dz in -1..=1 {
-            for dx in -1..=1 {
-                let mut c = Chunk::empty();
-                c.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-                c.smoothing = [1; CELLS_PER_CHUNK_AREA];
-                if dx == 0 && dz == 0 {
-                    // A 4x4 stone block on whole-cell heights — the cliff is a
-                    // material + cell-height break with zero subcell relief.
-                    for lz in 6..10 {
-                        for lx in 6..10 {
-                            let idx = (lz * EDGE + lx) as usize;
-                            c.underlay[idx] = Material::Stone;
-                            c.height[idx] = 512;
-                        }
-                    }
-                }
-                world.insert_chunk(ChunkPos { x: dx, z: dz }, c);
-            }
-        }
-        world
-    }
-
-    /// Twice the xz-projected area of a triangle — a wall face is exactly
-    /// vertical, so it projects to a line (zero area) while every cap
-    /// triangle keeps a footprint; the split separates the two without
-    /// guessing from heights.
-    fn xz_area_doubled(t: &DrawTriangle) -> f32 {
-        let [a, b, c] = &t.verts;
-        ((b.x - a.x) * (c.z - a.z) - (c.x - a.x) * (b.z - a.z)).abs()
-    }
-
-    #[test]
-    fn a_delta_silhouette_wall_coverage_is_complete() {
-        // The picket-fence regression from the causeway demo: gating marched
-        // segments on the window center's point level skips every window
-        // whose center lands on the low-material side — along a contour the
-        // center alternates sides subcell to subcell, halving the wall into
-        // isolated wedges. The side-aware per-segment gate must close the
-        // full perimeter: the emitted wall top-edge length covers the
-        // square's 8 m perimeter — the marched chamfers trade the corner
-        // right angles for diagonals and the clip-gap walls close the
-        // chamfer slivers' lattice edges, so the total sits just above 8 —
-        // far above the alternating half.
-        let world = delta_column_world();
-        let mesh = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        let total: f32 = mesh
-            .iter()
-            .filter(|t| y_span(t) > 0.5)
-            .filter_map(|t| {
-                // Each wall quad's top edge rides exactly one of its two
-                // triangles — the one holding both top vertices.
-                let tops: Vec<&Vertex> = t.verts.iter().filter(|v| v.y > 1.0).collect();
-                (tops.len() == 2).then(|| (tops[1].x - tops[0].x).hypot(tops[1].z - tops[0].z))
-            })
-            .sum();
-        assert!(
-            (7.5..8.6).contains(&total),
-            "wall top-edge coverage {total} m over an 8 m perimeter — alternating gaps halve it",
-        );
-    }
-
-    #[test]
-    fn a_material_block_wall_coverage_is_complete() {
-        // Tripwire: a relief-free material-boundary cliff must close its full
-        // perimeter. The bug lifted the marched wall top through the
-        // window-center owner rather than the anchored (high) side; on an
-        // axis-aligned boundary the owner sits on the low (grass) side for
-        // every boundary window, so every wall top collapses onto the grass
-        // plate and the whole face vanishes (zero-height, see-through). A 4x4
-        // cell stone block on flat grass has a 16 m perimeter; the emitted
-        // wall top-edge length must cover it, far above the collapsed
-        // near-zero the bug leaves.
-        let world = material_block_world();
-        let mesh = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        let block_top = 512.0 / 256.0;
-        let total: f32 = mesh
-            .iter()
-            .filter(|t| y_span(t) > 0.5)
-            .filter_map(|t| {
-                // Each wall quad's top edge rides the triangle holding both
-                // top (block-height) vertices.
-                let tops: Vec<&Vertex> = t.verts.iter().filter(|v| v.y > block_top - 0.5).collect();
-                (tops.len() == 2).then(|| (tops[1].x - tops[0].x).hypot(tops[1].z - tops[0].z))
-            })
-            .sum();
-        assert!(
-            (15.0..17.5).contains(&total),
-            "wall top-edge coverage {total} m over a 16 m perimeter — \
-             a relief-free material-boundary cliff collapsed its faces",
-        );
-    }
-
-    #[test]
-    fn caps_split_at_a_delta_break_without_draping() {
-        // The zigzag regression from the causeway demo: a contour vertex on
-        // the break line lifted through whichever subcell it floors into
-        // drapes the grass cap up the column flank and sags the stone cap
-        // down it with subcell parity. Split caps read their own side: every
-        // cap triangle (nonzero footprint — walls project to a line) lies
-        // wholly on its side's plate, the grass ground at zero, the stone
-        // column at its raised level, nothing in between.
-        let world = delta_column_world();
-        let mesh = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        let high = 300.0 / 256.0;
-        let (mut on_ground, mut on_column) = (false, false);
-        for t in &mesh {
-            if xz_area_doubled(t) < 1e-6 {
-                continue; // vertical wall faces own the gap
-            }
-            // Every cap triangle lies wholly on one plate: a draped triangle
-            // would mix the two levels (or hit an interpolated height between
-            // them) within one footprint.
-            let ground = t.verts.iter().all(|v| v.y.abs() < 1e-3);
-            let column = t.verts.iter().all(|v| (v.y - high).abs() < 1e-3);
-            assert!(
-                ground || column,
-                "cap triangle at ({}, {}) spans heights {:?} — a drape across the break",
-                t.verts[0].x,
-                t.verts[0].z,
-                t.verts.map(|v| v.y),
-            );
-            on_ground |= ground;
-            on_column |= column;
-        }
-        assert!(
-            on_ground && on_column,
-            "both caps are present (ground {on_ground}, column {on_column})",
-        );
-    }
-
-    #[test]
-    fn delta_break_walls_seal_both_caps() {
-        // Watertightness across the delta break: every wall top vertex
-        // coincides with a high-cap vertex, and every wall bottom vertex
-        // coincides with a low-cap vertex at the same position — under
-        // closure the base is the low cap's committed edge exactly, so the
-        // wall spans precisely the vertical gap the split caps open, no
-        // pinholes, no overlap.
-        let world = delta_column_world();
-        let mesh = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        let cap_verts: Vec<&Vertex> = mesh
-            .iter()
-            .filter(|t| xz_area_doubled(t) > 1e-6)
-            .flat_map(|t| t.verts.iter())
-            .collect();
-        let seals = |x: f32, z: f32, y: f32| {
-            cap_verts
-                .iter()
-                .any(|c| (c.x - x).abs() < 1e-5 && (c.z - z).abs() < 1e-5 && (c.y - y).abs() < 1e-4)
-        };
-        let walls: Vec<&DrawTriangle> = mesh.iter().filter(|t| y_span(t) > 0.5).collect();
-        assert!(!walls.is_empty(), "the break carries walls");
-        for v in walls.iter().flat_map(|t| t.verts.iter()) {
-            if v.y > 1.0 {
-                assert!(
-                    seals(v.x, v.z, v.y),
-                    "wall top ({}, {}, {}) misses the high cap",
-                    v.x,
-                    v.z,
-                    v.y,
-                );
-            } else {
-                assert!(
-                    seals(v.x, v.z, v.y),
-                    "wall base ({}, {}, {}) misses the low cap",
-                    v.x,
-                    v.z,
-                    v.y,
-                );
-            }
-        }
-    }
-
-    /// The cap-split locations of a mesh: quantized xz positions where cap
-    /// vertices (nonzero-footprint triangles) sit on two plates more than
-    /// half a meter apart — the places the caps genuinely split over a
-    /// break and a wall must span.
-    fn cap_splits(mesh: &[DrawTriangle]) -> Vec<(f32, f32, f32, f32)> {
-        let mut heights: BTreeMap<(i64, i64), (f32, f32)> = BTreeMap::new();
-        for v in mesh
-            .iter()
-            .filter(|t| xz_area_doubled(t) > 1e-6)
-            .flat_map(|t| t.verts.iter())
-        {
-            let key = ((v.x * 256.0).round() as i64, (v.z * 256.0).round() as i64);
-            let entry = heights.entry(key).or_insert((v.y, v.y));
-            entry.0 = entry.0.min(v.y);
-            entry.1 = entry.1.max(v.y);
-        }
-        heights
-            .into_iter()
-            .filter(|&(_, (lo, hi))| hi - lo > 0.5)
-            .map(|((qx, qz), (lo, hi))| (qx as f32 / 256.0, qz as f32 / 256.0, lo, hi))
-            .collect()
-    }
-
-    /// Shared assertion for authored-plateau fixtures — the live causeway
-    /// probe's invariant, in-repo: every cap triangle (nonzero footprint)
-    /// lies level on one plateau (a triangle bridging two plateaus is the
-    /// slanted "plane not sitting flat" the demo showed — a flap fan
-    /// chording across a break), the authored breaks genuinely split the
-    /// caps, and every split is spanned by a wall at exactly its two
-    /// plates.
-    fn assert_plateaus_flat_and_sealed(world: &World, at: ChunkPos) {
-        let mesh = mesh_chunk(world, at, &StyleTable::default());
-        for t in &mesh {
-            if xz_area_doubled(t) < 1e-6 {
-                continue; // vertical wall faces
-            }
-            let ys = t.verts.map(|v| v.y);
-            assert!(
-                (ys[0] - ys[1]).abs() < 1e-3 && (ys[0] - ys[2]).abs() < 1e-3,
-                "cap triangle at ({}, {}) bridges plateaus: {ys:?}",
-                t.verts[0].x,
-                t.verts[0].z,
-            );
-        }
-        let splits = cap_splits(&mesh);
-        assert!(!splits.is_empty(), "the authored breaks split the caps");
-        let walls: Vec<&DrawTriangle> = mesh
-            .iter()
-            .filter(|t| xz_area_doubled(t) < 1e-6 && y_span(t) > 0.25)
-            .collect();
-        // A wall covers `(x, z)` at height `y` when the point lies on some
-        // wall triangle's edge — walls span whole subcell edges and marched
-        // segments, so a split can sit mid-edge (a T-junction on the same
-        // straight line), not only on a wall vertex.
-        let covered = |x: f32, z: f32, y: f32| {
-            walls.iter().any(|t| {
-                (0..3).any(|k| {
-                    let a = &t.verts[k];
-                    let b = &t.verts[(k + 1) % 3];
-                    let (ex, ez) = (b.x - a.x, b.z - a.z);
-                    let len_sq = ex * ex + ez * ez;
-                    if len_sq < 1e-8 {
-                        return false;
-                    }
-                    let (px, pz) = (x - a.x, z - a.z);
-                    if (ex * pz - ez * px).abs() > 1e-3 {
-                        return false; // off the edge's xz line
-                    }
-                    let t_on = (px * ex + pz * ez) / len_sq;
-                    if !(-1e-4..=1.0 + 1e-4).contains(&t_on) {
-                        return false;
-                    }
-                    (a.y + (b.y - a.y) * t_on - y).abs() < 1e-3
-                })
-            })
-        };
-        for (x, z, lo, hi) in splits {
-            assert!(
-                covered(x, z, hi),
-                "split at ({x}, {z}) has no wall top at {hi}",
-            );
-            assert!(
-                covered(x, z, lo),
-                "split at ({x}, {z}) has no wall base at {lo}",
-            );
-        }
-    }
-
-    #[test]
-    fn a_three_plateau_junction_stays_flat_and_sealed() {
-        // Three same-material plateaus (deltas 0 / 128 / 256) meeting at a
-        // point inside a stone square on grass, smoothing on — the junction
-        // regime the causeway probe caught: mixed windows near the zone
-        // lines hold same-label corner samples on different plateaus, and a
-        // whole-window flap fan chords across the break (pairs like
-        // 0.5 ↔ 1.0 in one footprint triangle). The clipped flaps must keep
-        // every cap triangle on one plateau and wall every split.
-        let mut world = World::new();
-        world.insert_smoothing_profile(
-            1,
-            SmoothingProfile {
-                iterations: 2,
-                degrees: 45,
-            },
-        );
-        for dz in -1..=1 {
-            for dx in -1..=1 {
-                let mut c = Chunk::empty();
-                c.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-                c.smoothing = [1; CELLS_PER_CHUNK_AREA];
-                world.insert_chunk(ChunkPos { x: dx, z: dz }, c);
-            }
-        }
-        let sub = SUB as usize;
-        let body_lo = sub4(28);
-        let body_hi = sub4(44);
-        let split = sub4(36);
-        for lz in 0..EDGE {
-            for lx in 0..EDGE {
-                let mut points = [UNDERLAY_POINT_INHERIT; SUBCELLS_PER_CELL];
-                let mut deltas = [0i16; SUBCELLS_PER_CELL];
-                let mut any = false;
-                for sz in 0..sub {
-                    for sx in 0..sub {
-                        let gx = lx * SUB + sx as i32;
-                        let gz = lz * SUB + sz as i32;
-                        if (body_lo..body_hi).contains(&gx) && (body_lo..body_hi).contains(&gz) {
-                            points[sz * sub + sx] = Material::Stone.to_u8();
-                            // Zone lines x = 9 m and z = 9 m meet at the
-                            // junction and run out through the perimeter.
-                            deltas[sz * sub + sx] = if gx < split {
-                                0
-                            } else if gz < split {
-                                128
-                            } else {
-                                256
-                            };
-                            any = true;
-                        }
-                    }
-                }
-                if any {
-                    let cell = CellPos { x: lx, z: lz };
-                    world.set_cell_points(cell, &points);
-                    world.set_cell_heights(cell, &deltas);
-                }
-            }
-        }
-        assert_plateaus_flat_and_sealed(&world, ChunkPos { x: 0, z: 0 });
-    }
-
-    #[test]
-    fn a_two_height_silhouette_against_grass_stays_flat_and_sealed() {
-        // An L-shaped stone region on grass whose two arms stand at
-        // different heights (deltas 128 / 256), smoothing on — the internal
-        // same-material break line runs out through the stone/grass
-        // perimeter, the second junction regime from the causeway probe.
-        // Clipped flaps keep every cap triangle level; the walls close both
-        // the perimeter and the internal break.
-        let mut world = World::new();
-        world.insert_smoothing_profile(
-            1,
-            SmoothingProfile {
-                iterations: 2,
-                degrees: 45,
-            },
-        );
-        for dz in -1..=1 {
-            for dx in -1..=1 {
-                let mut c = Chunk::empty();
-                c.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-                c.smoothing = [1; CELLS_PER_CHUNK_AREA];
-                world.insert_chunk(ChunkPos { x: dx, z: dz }, c);
-            }
-        }
-        let sub = SUB as usize;
-        let arm_x = sub4(30)..sub4(36);
-        let arm_z = sub4(26)..sub4(42);
-        let foot_x = sub4(36)..sub4(42);
-        let foot_z = sub4(34)..sub4(42);
-        for lz in 0..EDGE {
-            for lx in 0..EDGE {
-                let mut points = [UNDERLAY_POINT_INHERIT; SUBCELLS_PER_CELL];
-                let mut deltas = [0i16; SUBCELLS_PER_CELL];
-                let mut any = false;
-                for sz in 0..sub {
-                    for sx in 0..sub {
-                        let gx = lx * SUB + sx as i32;
-                        let gz = lz * SUB + sz as i32;
-                        // Vertical arm at +128; the foot at +256.
-                        let arm = arm_x.contains(&gx) && arm_z.contains(&gz);
-                        let foot = foot_x.contains(&gx) && foot_z.contains(&gz);
-                        if arm || foot {
-                            points[sz * sub + sx] = Material::Stone.to_u8();
-                            deltas[sz * sub + sx] = if arm { 128 } else { 256 };
-                            any = true;
-                        }
-                    }
-                }
-                if any {
-                    let cell = CellPos { x: lx, z: lz };
-                    world.set_cell_points(cell, &points);
-                    world.set_cell_heights(cell, &deltas);
-                }
-            }
-        }
-        assert_plateaus_flat_and_sealed(&world, ChunkPos { x: 0, z: 0 });
-    }
-
-    #[test]
-    fn zeroed_deltas_mesh_identically() {
-        // Tripwire: a world whose height deltas are all zero meshes byte-for-
-        // byte as one that never carried the plane. Author then clear a cell's
-        // deltas over a varied world; the net-zero relief must collapse back
-        // to the cell-stride caps and walls, adding and moving nothing.
-        let at = ChunkPos { x: 0, z: 0 };
-        let mut world = World::new();
-        let mut chunk = Chunk::empty();
-        for lz in 0..EDGE {
-            for lx in 0..EDGE {
-                let i = (lz * EDGE + lx) as usize;
-                chunk.underlay[i] = if lx < 8 {
-                    Material::Grass
-                } else {
-                    Material::Sand
-                };
-                chunk.height[i] = 8 * lx; // gentle ramp, no cliffs
-            }
-        }
-        world.insert_chunk(at, chunk);
-        let baseline = mesh_chunk(&world, at, &StyleTable::default());
-
-        let mut authored = world.clone();
-        authored.set_cell_heights(CellPos { x: 4, z: 4 }, &[50; SUBCELLS_PER_CELL]);
-        authored.set_cell_heights(CellPos { x: 4, z: 4 }, &[]); // clears to zero
-        let after = mesh_chunk(&authored, at, &StyleTable::default());
-
-        assert_eq!(
-            baseline, after,
-            "a net-zero delta plane meshes identically to no plane",
-        );
-    }
-
-    #[test]
-    fn a_fillable_void_joint_floors_and_walls() {
-        // Tripwire: an enclosed Void well one meter deep in a grass field
-        // closes as a real flat-bottomed cut — a floor cap at the void
-        // points' stored height and rim walls from the grass down to it, no
-        // open bottom. A prediction pass blind to the void depth would leave
-        // the groove open (a wall to a fixed drop, or no wall at all); the
-        // floor and its walls both read the one total height plane.
-        let mut world = World::new();
-        let mut chunk = Chunk::empty();
-        chunk.underlay = [Material::Grass; CELLS_PER_CHUNK_AREA];
-        world.insert_chunk(ChunkPos { x: 0, z: 0 }, chunk);
-        // Cell (7, 7) is a full Void well, floored one meter down and enclosed
-        // by grass on every side.
-        let cell = CellPos { x: 7, z: 7 };
-        world.set_cell_points(cell, &[Material::Void.to_u8(); SUBCELLS_PER_CELL]);
-        world.set_cell_heights(cell, &[-256; SUBCELLS_PER_CELL]);
-
-        let mesh = mesh_chunk(&world, ChunkPos { x: 0, z: 0 }, &StyleTable::default());
-        let floor = -256.0 / OCTIMETERS_PER_METER;
-        let has_floor = mesh.iter().any(|t| {
-            xz_area_doubled(t) > 1e-6 && t.verts.iter().all(|v| (v.y - floor).abs() < 1e-3)
-        });
-        assert!(
-            has_floor,
-            "the enclosed void floors over at its stored depth"
-        );
-        // A rim wall spans the grass edge (y ~ 0) down to the floor (y ~ -1 m)
-        // — the groove is closed, not open-bottomed.
-        let has_rim_wall = mesh.iter().any(|t| {
-            xz_area_doubled(t) < 1e-6
-                && t.verts.iter().any(|v| v.y > -0.1)
-                && t.verts.iter().any(|v| (v.y - floor).abs() < 1e-3)
-        });
-        assert!(
-            has_rim_wall,
-            "the void rim walls down to the floor, no open bottom",
-        );
-    }
 }
